Modularize GML runtime and transpiler
Refactor the monolithic GML runtime into a modular package: move the large GML_RUNTIME_SCRIPT out of src/conversion/gml_runtime.py into src/conversion/gml_runtime_parts/segments/*.gd and assemble it via src/conversion/gml_runtime_parts/script.py. Add a writer module and __init__ for the new package, then update gml_runtime.py to import the assembled script, resource/relative path constants and the write_gml_runtime function and re-export them. Also add a suite of new src/conversion/gml_transpiler_parts/* modules to break up the transpiler into smaller components for improved maintainability and readability.
