+++ src/conversion/gml_transpiler.py
@@ -1,4569 +1,77 @@
+# pyright: reportPrivateUsage=false
 from __future__ import annotations
 
-import json
-import hashlib
-from dataclasses import dataclass
-from typing import Iterable, Literal, Mapping, MutableMapping, MutableSet, TypeAlias
+from src.conversion.gml_transpiler_parts.api import transpile_gml_code
+from src.conversion.gml_transpiler_parts.constants import _BUILTIN_VARIABLE_REGISTRY
+from src.conversion.gml_transpiler_parts.expression_parser import (
+    _ExpressionParser,
+    _parse_gml_expression,
+)
+from src.conversion.gml_transpiler_parts.expression_service import (
+    transpile_gml_condition,
+    transpile_gml_expression,
+)
+from src.conversion.gml_transpiler_parts.model import (
+    GMLTranspileError,
+    _ArrayLiteral,
+    _Binary,
+    _BuiltinVariableMetadata,
+    _Call,
+    _DSMapAccess,
+    _Expression,
+    _FunctionLiteral,
+    _FunctionParameter,
+    _Grouped,
+    _Index,
+    _Literal,
+    _Member,
+    _Name,
+    _NameOf,
+    _NewCall,
+    _NumberLiteral,
+    _ScopeContext,
+    _StaticDeclaration,
+    _StringLiteral,
+    _StructAccess,
+    _StructLiteral,
+    _Ternary,
+    _Token,
+    _Unary,
+)
+from src.conversion.gml_transpiler_parts.tokens import _expression_tokens, _tokenize
 
-
-_AssignmentOperator: TypeAlias = Literal[
-    "??=",
-    "<<=",
-    ">>=",
-    ":=",
-    "+=",
-    "-=",
-    "*=",
-    "/=",
-    "%=",
-    "&=",
-    "|=",
-    "^=",
-    "=",
+__all__ = [
+    "GMLTranspileError",
+    "_ArrayLiteral",
+    "_Binary",
+    "_BuiltinVariableMetadata",
+    "_BUILTIN_VARIABLE_REGISTRY",
+    "_Call",
+    "_DSMapAccess",
+    "_Expression",
+    "_ExpressionParser",
+    "_FunctionLiteral",
+    "_FunctionParameter",
+    "_Grouped",
+    "_Index",
+    "_Literal",
+    "_Member",
+    "_Name",
+    "_NameOf",
+    "_NewCall",
+    "_NumberLiteral",
+    "_ScopeContext",
+    "_StaticDeclaration",
+    "_StringLiteral",
+    "_StructAccess",
+    "_StructLiteral",
+    "_Ternary",
+    "_Token",
+    "_Unary",
+    "_expression_tokens",
+    "_parse_gml_expression",
+    "_tokenize",
+    "transpile_gml_code",
+    "transpile_gml_condition",
+    "transpile_gml_expression",
 ]
-
-_IncrementDelta: TypeAlias = Literal[-1, 1]
-_IncrementMode: TypeAlias = Literal["prefix", "postfix"]
-
-
-class GMLTranspileError(ValueError):
-    """Raised when the small GML subset transpiler cannot parse input."""
-
-
-@dataclass(frozen=True)
-class _Token:
-    kind: str
-    value: str
-
-
-@dataclass(frozen=True)
-class _BuiltinVariableMetadata:
-    scope: str
-    default: str
-    mutable: bool
-    is_array: bool
-    subsystem: str
-
-
-@dataclass(frozen=True)
-class _ScopeContext:
-    self_expression: str = "self"
-    other_expression: str = "other"
-    instance_target: str | None = None
-    global_scope: bool = False
-    global_names: frozenset[str] = frozenset()
-    asset_names: frozenset[str] = frozenset()
-    static_scope: str | None = None
-    static_names: frozenset[str] = frozenset()
-    static_prefix: str = "gml_static"
-
-
-_DEFAULT_SCOPE_CONTEXT = _ScopeContext()
-
-
-@dataclass(frozen=True)
-class _Name:
-    value: str
-
-
-@dataclass(frozen=True)
-class _NameOf:
-    value: str
-
-
-@dataclass(frozen=True)
-class _Literal:
-    value: str
-
-
-@dataclass(frozen=True)
-class _StringLiteral:
-    value: str
-
-
-@dataclass(frozen=True)
-class _NumberLiteral:
-    value: str
-    is_float_like: bool
-
-
-@dataclass(frozen=True)
-class _Unary:
-    operator: str
-    operand: _Expression
-
-
-@dataclass(frozen=True)
-class _Binary:
-    left: _Expression
-    operator: str
-    right: _Expression
-
-
-@dataclass(frozen=True)
-class _Ternary:
-    condition: _Expression
-    true_expr: _Expression
-    false_expr: _Expression
-
-
-@dataclass(frozen=True)
-class _Call:
-    callee: _Expression
-    args: tuple[_Expression, ...]
-
-
-@dataclass(frozen=True)
-class _ArrayLiteral:
-    elements: tuple[_Expression, ...]
-
-
-@dataclass(frozen=True)
-class _FunctionParameter:
-    name: str
-    default: _Expression | None
-
-
-@dataclass(frozen=True)
-class _StaticDeclaration:
-    name: str
-    value_source: str
-
-
-@dataclass(frozen=True)
-class _FunctionLiteral:
-    name: str | None
-    parameters: tuple[_FunctionParameter, ...]
-    body_lines: tuple[str, ...]
-    is_constructor: bool = False
-    static_scope_id: str | None = None
-
-
-@dataclass(frozen=True)
-class _NewCall:
-    constructor: _Expression
-    args: tuple[_Expression, ...]
-
-
-@dataclass(frozen=True)
-class _StructLiteral:
-    fields: tuple[tuple[str, _Expression], ...]
-
-
-@dataclass(frozen=True)
-class _Index:
-    target: _Expression
-    index: _Expression
-
-
-@dataclass(frozen=True)
-class _StructAccess:
-    target: _Expression
-    key: _Expression
-
-
-@dataclass(frozen=True)
-class _DSMapAccess:
-    target: _Expression
-    key: _Expression
-
-
-@dataclass(frozen=True)
-class _Member:
-    target: _Expression
-    member: str
-
-
-@dataclass(frozen=True)
-class _Grouped:
-    expr: _Expression
-
-
-_Expression: TypeAlias = (
-    _Name
-    | _NameOf
-    | _Literal
-    | _StringLiteral
-    | _NumberLiteral
-    | _Unary
-    | _Binary
-    | _Ternary
-    | _Call
-    | _ArrayLiteral
-    | _FunctionLiteral
-    | _NewCall
-    | _StructLiteral
-    | _Index
-    | _StructAccess
-    | _DSMapAccess
-    | _Member
-    | _Grouped
-)
-
-
-_EOF = _Token("EOF", "")
-
-_MULTI_CHAR_OPERATORS = (
-    "??=",
-    "<<=",
-    ">>=",
-    ":=",
-    "??",
-    "<=",
-    ">=",
-    "==",
-    "!=",
-    "&&",
-    "||",
-    "^^",
-    "++",
-    "--",
-    "+=",
-    "-=",
-    "*=",
-    "/=",
-    "%=",
-    "&=",
-    "|=",
-    "^=",
-    "<<",
-    ">>",
-)
-
-_ASSIGNMENT_OPERATORS: tuple[_AssignmentOperator, ...] = (
-    "??=",
-    "<<=",
-    ">>=",
-    ":=",
-    "+=",
-    "-=",
-    "*=",
-    "/=",
-    "%=",
-    "&=",
-    "|=",
-    "^=",
-    "=",
-)
-
-_BINARY_PRECEDENCE = {
-    "??": 10,
-    "or": 20,
-    "||": 20,
-    "^^": 20,
-    "and": 30,
-    "&&": 30,
-    "|": 40,
-    "^": 50,
-    "&": 60,
-    "=": 70,
-    "==": 70,
-    "!=": 70,
-    "<": 70,
-    "<=": 70,
-    ">": 70,
-    ">=": 70,
-    "<<": 80,
-    ">>": 80,
-    "+": 90,
-    "-": 90,
-    "*": 100,
-    "/": 100,
-    "%": 100,
-    "div": 100,
-    "mod": 100,
-}
-
-_UNARY_PRECEDENCE = 110
-_POSTFIX_PRECEDENCE = 120
-_PRIMARY_PRECEDENCE = 130
-_TERNARY_PRECEDENCE = 5
-_GML_IDENTIFIER_MAX_LENGTH = 64
-_GENERATED_IDENTIFIER_PREFIX = "_gml_"
-
-_RIGHT_ASSOCIATIVE = {"??"}
-
-_GDSCRIPT_RESERVED_IDENTIFIERS = frozenset({
-    "as",
-    "break",
-    "class",
-    "class_name",
-    "const",
-    "continue",
-    "elif",
-    "else",
-    "enum",
-    "extends",
-    "false",
-    "for",
-    "func",
-    "if",
-    "in",
-    "is",
-    "match",
-    "null",
-    "pass",
-    "return",
-    "self",
-    "signal",
-    "static",
-    "super",
-    "true",
-    "var",
-    "void",
-    "when",
-    "while",
-})
-_GML_LITERAL_IDENTIFIERS = frozenset({
-    "false",
-    "null",
-    "self",
-    "super",
-    "true",
-})
-_GML_BUILTIN_CONSTANT_IDENTIFIERS = frozenset({
-    "NaN",
-    "all",
-    "false",
-    "infinity",
-    "nan",
-    "noone",
-    "null",
-    "pi",
-    "pointer_invalid",
-    "pointer_null",
-    "true",
-    "undefined",
-})
-_DIRECT_MEMBER_TARGETS = frozenset({
-    "global",
-    "other",
-    "self",
-    "super",
-})
-
-_BOOLEAN_RESULT_BINARY_OPERATORS = frozenset({
-    "&&",
-    "||",
-    "^^",
-    "and",
-    "or",
-    "=",
-    "==",
-    "!=",
-    "<",
-    "<=",
-    ">",
-    ">=",
-})
-
-_BOOLEAN_RESULT_FUNCTIONS = frozenset({
-    "bool",
-    "is_array",
-    "is_bool",
-    "is_callable",
-    "is_handle",
-    "is_infinity",
-    "is_int32",
-    "is_int64",
-    "is_method",
-    "is_nan",
-    "is_numeric",
-    "is_ptr",
-    "is_real",
-    "is_string",
-    "is_struct",
-    "struct_exists",
-    "is_undefined",
-    "keyboard_check",
-})
-
-_ARITHMETIC_RUNTIME_FUNCTIONS = {
-    "+": "gml_add",
-    "-": "gml_sub",
-    "*": "gml_mul",
-    "%": "gml_mod",
-    "mod": "gml_mod",
-}
-
-_BITWISE_RUNTIME_FUNCTIONS = {
-    "&": "gml_bit_and",
-    "|": "gml_bit_or",
-    "^": "gml_bit_xor",
-    "<<": "gml_shift_left",
-    ">>": "gml_shift_right",
-}
-
-_COMPOUND_RUNTIME_FUNCTIONS: dict[_AssignmentOperator, str] = {
-    "&=": "gml_bit_and",
-    "+=": "gml_add",
-    "-=": "gml_sub",
-    "*=": "gml_mul",
-    "/=": "gml_div",
-    "%=": "gml_mod",
-    "<<=": "gml_shift_left",
-    ">>=": "gml_shift_right",
-    "|=": "gml_bit_or",
-    "^=": "gml_bit_xor",
-}
-
-_OPERATOR_REPLACEMENTS = {
-    "&&": "and",
-    "||": "or",
-    "=": "==",
-    "mod": "%",
-}
-
-_NAME_REPLACEMENTS = {
-    "all": "GMRuntime.gml_instance_all()",
-    "global": "GMRuntime.gml_global_scope()",
-    "infinity": "INF",
-    "NaN": "NAN",
-    "nan": "NAN",
-    "noone": "GMRuntime.gml_instance_noone()",
-    "pi": "PI",
-    "pointer_invalid": "GMRuntime.gml_pointer_invalid()",
-    "pointer_null": "GMRuntime.gml_pointer_null()",
-    "undefined": "GMRuntime.gml_undefined()",
-}
-
-_BLOCK_DELIMITER_REPLACEMENTS = {
-    "begin": "{",
-    "end": "}",
-}
-
-_INSTANCE_NAME_REPLACEMENTS = {
-    "x": "position.x",
-    "y": "position.y",
-}
-
-_LEGACY_GLOBAL_BUILTINS = frozenset({"health", "lives", "score"})
-
-_BUILTIN_VARIABLE_REGISTRY = {
-    "argument": _BuiltinVariableMetadata("global", "[]", False, False, "script_arguments"),
-    "argument_count": _BuiltinVariableMetadata("global", "0", False, False, "script_arguments"),
-    "async_load": _BuiltinVariableMetadata("global", "{}", False, False, "async_event"),
-    "bbox_bottom": _BuiltinVariableMetadata("instance", "0", False, False, "collision_bounds"),
-    "bbox_left": _BuiltinVariableMetadata("instance", "0", False, False, "collision_bounds"),
-    "bbox_right": _BuiltinVariableMetadata("instance", "0", False, False, "collision_bounds"),
-    "bbox_top": _BuiltinVariableMetadata("instance", "0", False, False, "collision_bounds"),
-    "current_time": _BuiltinVariableMetadata("global", "0", False, False, "time"),
-    "depth": _BuiltinVariableMetadata("instance", "0", True, False, "rendering"),
-    "direction": _BuiltinVariableMetadata("instance", "0", True, False, "motion"),
-    "event_data": _BuiltinVariableMetadata("global", "{}", False, False, "event"),
-    "fps": _BuiltinVariableMetadata("global", "0", False, False, "time"),
-    "id": _BuiltinVariableMetadata("instance", "undefined", False, False, "identity"),
-    "image_index": _BuiltinVariableMetadata("instance", "0", True, False, "sprite"),
-    "image_number": _BuiltinVariableMetadata("instance", "0", False, False, "sprite"),
-    "instance_count": _BuiltinVariableMetadata("global", "0", False, False, "instances"),
-    "layer": _BuiltinVariableMetadata("instance", "0", True, False, "rendering"),
-    "object_index": _BuiltinVariableMetadata("instance", "undefined", False, False, "identity"),
-    "path_index": _BuiltinVariableMetadata("instance", "undefined", True, False, "path"),
-    "path_position": _BuiltinVariableMetadata("instance", "0", True, False, "path"),
-    "path_scale": _BuiltinVariableMetadata("instance", "1", True, False, "path"),
-    "path_speed": _BuiltinVariableMetadata("instance", "0", True, False, "path"),
-    "room": _BuiltinVariableMetadata("global", "undefined", False, False, "room"),
-    "room_height": _BuiltinVariableMetadata("global", "0", False, False, "room"),
-    "room_width": _BuiltinVariableMetadata("global", "0", False, False, "room"),
-    "speed": _BuiltinVariableMetadata("instance", "0", True, False, "motion"),
-    "sprite_index": _BuiltinVariableMetadata("instance", "undefined", True, False, "sprite"),
-    "view_angle": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_camera": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_current": _BuiltinVariableMetadata("global", "undefined", False, True, "view"),
-    "view_enabled": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_hborder": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_hport": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_hspeed": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_hview": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_object": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_surface_id": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_vborder": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_visible": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_vspeed": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_wport": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_wview": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_xport": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_xview": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_yport": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "view_yview": _BuiltinVariableMetadata("global", "undefined", True, True, "view"),
-    "visible": _BuiltinVariableMetadata("instance", "true", True, False, "rendering"),
-    "x": _BuiltinVariableMetadata("instance", "0", True, False, "transform"),
-    "y": _BuiltinVariableMetadata("instance", "0", True, False, "transform"),
-}
-
-_BUILTIN_GLOBAL_VARIABLES = frozenset(
-    name for name, metadata in _BUILTIN_VARIABLE_REGISTRY.items()
-    if metadata.scope == "global" and not metadata.is_array
-)
-_BUILTIN_ARRAY_VARIABLES = frozenset(
-    name for name, metadata in _BUILTIN_VARIABLE_REGISTRY.items()
-    if metadata.is_array
-)
-_READ_ONLY_BUILTIN_VARIABLES = frozenset(
-    name for name, metadata in _BUILTIN_VARIABLE_REGISTRY.items()
-    if not metadata.mutable
-)
-_BUILTIN_INSTANCE_VARIABLES = frozenset(_BUILTIN_VARIABLE_REGISTRY)
-
-_VIRTUAL_KEY_ACTIONS = {
-    "vk_left": "ui_left",
-    "vk_right": "ui_right",
-    "vk_up": "ui_up",
-    "vk_down": "ui_down",
-}
-
-_VIRTUAL_KEY_CONSTANTS = {
-    "vk_shift": "KEY_SHIFT",
-}
-
-_RUNTIME_FUNCTIONS = {
-    "int64": "gml_int64",
-    "is_array": "is_array",
-    "is_bool": "is_bool",
-    "is_callable": "is_callable",
-    "is_handle": "is_handle",
-    "is_infinity": "is_infinity",
-    "is_int32": "is_int32",
-    "is_int64": "is_int64",
-    "is_method": "is_method",
-    "is_nan": "is_nan_value",
-    "is_numeric": "is_numeric",
-    "is_ptr": "is_ptr",
-    "is_real": "is_real",
-    "is_string": "is_string",
-    "is_struct": "is_struct",
-    "is_undefined": "is_undefined",
-    "handle_parse": "gml_handle_parse",
-    "method_get_index": "gml_method_get_index",
-    "method_get_self": "gml_method_get_self",
-    "real": "gml_real",
-    "ptr": "gml_ptr",
-    "sqrt": "gml_sqrt",
-    "typeof": "gml_typeof",
-    "with_targets": "gml_with_targets",
-    "string": "gml_string",
-    "bool": "gml_bool",
-}
-
-_STRUCT_RUNTIME_FUNCTIONS = {
-    "struct_exists": "gml_struct_exists",
-    "struct_get": "gml_struct_get",
-    "struct_get_names": "gml_struct_get_names",
-    "struct_names_count": "gml_struct_names_count",
-    "struct_set": "gml_struct_set",
-    "struct_remove": "gml_struct_remove",
-    "struct_foreach": "gml_struct_foreach",
-    "static_get": "gml_static_get",
-    "static_set": "gml_static_set",
-    "is_instanceof": "gml_is_instanceof",
-    "instanceof": "gml_instanceof",
-    "variable_get_hash": "gml_variable_get_hash",
-    "struct_get_from_hash": "gml_struct_get_from_hash",
-    "struct_set_from_hash": "gml_struct_set_from_hash",
-    "struct_exists_from_hash": "gml_struct_exists_from_hash",
-    "struct_remove_from_hash": "gml_struct_remove_from_hash",
-}
-
-_VARIABLE_RUNTIME_FUNCTIONS = {
-    "method_call": "gml_method_call",
-    "method": "gml_method",
-    "ref_create": "gml_ref_create",
-    "variable_clone": "gml_variable_clone",
-    "variable_instance_exists": "gml_variable_instance_exists",
-    "variable_instance_get": "gml_variable_instance_get",
-    "variable_instance_set": "gml_variable_instance_set",
-    "variable_instance_get_names": "gml_variable_instance_get_names",
-    "variable_instance_names_count": "gml_variable_instance_names_count",
-    "variable_global_exists": "gml_variable_global_exists",
-    "variable_global_get": "gml_variable_global_get",
-    "variable_global_set": "gml_variable_global_set",
-    "variable_struct_get": "gml_variable_struct_get",
-}
-
-_DS_MAP_RUNTIME_FUNCTIONS = {
-    "ds_map_exists": "gml_ds_map_exists",
-    "ds_map_find_value": "gml_ds_map_find_value",
-}
-
-_ARRAY_RUNTIME_FUNCTIONS = {
-    "array_equals": "gml_array_equals",
-    "array_push": "gml_array_push",
-}
-
-
-def transpile_gml_expression(
-    source: str,
-    local_names: Iterable[str] | None = None,
-    enum_values: MutableMapping[str, dict[str, int]] | None = None,
-    enum_names: Iterable[str] | None = None,
-    scope_context: _ScopeContext | None = None,
-    macro_values: Mapping[str, str] | None = None,
-    global_names: Iterable[str] | None = None,
-    asset_names: Iterable[str] | None = None,
-) -> str:
-    """Transpile a single GML expression to a GDScript expression."""
-    scope_context = _scope_context_with_global_names(
-        _normalize_scope_context(scope_context),
-        global_names,
-        asset_names=asset_names,
-    )
-    parser = _ExpressionParser(
-        _expression_tokens(source),
-        enum_values=enum_values,
-        enum_names=enum_names,
-        scope_context=scope_context,
-        macro_values=macro_values,
-    )
-    expr = parser.parse()
-    _reject_enum_mutation_expression(expr, enum_names)
-    return _emit_expression(
-        expr,
-        _normalize_local_names(local_names),
-        scope_context=scope_context,
-    )[0]
-
-
-def transpile_gml_condition(
-    source: str,
-    local_names: Iterable[str] | None = None,
-    enum_values: MutableMapping[str, dict[str, int]] | None = None,
-    enum_names: Iterable[str] | None = None,
-    scope_context: _ScopeContext | None = None,
-    macro_values: Mapping[str, str] | None = None,
-    global_names: Iterable[str] | None = None,
-    asset_names: Iterable[str] | None = None,
-) -> str:
-    """Transpile a GML condition using GameMaker truthiness semantics."""
-    scope_context = _scope_context_with_global_names(
-        _normalize_scope_context(scope_context),
-        global_names,
-        asset_names=asset_names,
-    )
-    parser = _ExpressionParser(
-        _expression_tokens(source),
-        enum_values=enum_values,
-        enum_names=enum_names,
-        scope_context=scope_context,
-        macro_values=macro_values,
-    )
-    expr = parser.parse()
-    _reject_enum_mutation_expression(expr, enum_names)
-    return _emit_truthy_expression(
-        expr,
-        _normalize_local_names(local_names),
-        scope_context=scope_context,
-    )
-
-
-def _parse_gml_expression(
-    source: str,
-    enum_values: MutableMapping[str, dict[str, int]] | None = None,
-    enum_names: Iterable[str] | None = None,
-    macro_values: Mapping[str, str] | None = None,
-    macro_expansion_stack: frozenset[str] | None = None,
-    scope_context: _ScopeContext | None = None,
-) -> _Expression:
-    parser = _ExpressionParser(
-        _expression_tokens(source),
-        enum_values=enum_values,
-        enum_names=enum_names,
-        scope_context=scope_context,
-        macro_values=macro_values,
-        macro_expansion_stack=macro_expansion_stack,
-    )
-    return parser.parse()
-
-
-def transpile_gml_code(
-    source: str,
-    indent: str = "\t",
-    instance_variables: MutableSet[str] | None = None,
-    inherited_event_call: str | None = None,
-    macro_configuration: str | None = None,
-    top_level_global_scope: bool = False,
-    legacy_global_builtins: bool = False,
-    asset_names: Iterable[str] | None = None,
-    static_scope_prefix: str | None = None,
-) -> str:
-    """Transpile supported GML statements to GDScript."""
-    parser = _StatementParser(
-        _tokenize(_join_macro_continuation_lines(_strip_comments(source))),
-        instance_variables=instance_variables,
-        inherited_event_call=inherited_event_call,
-        macro_configuration=macro_configuration,
-        top_level_global_scope=top_level_global_scope,
-        global_names=_LEGACY_GLOBAL_BUILTINS if legacy_global_builtins else None,
-        asset_names=asset_names,
-        static_scope_prefix=static_scope_prefix,
-    )
-    lines = parser.parse()
-
-    if not lines:
-        return f"{indent}pass"
-
-    return "\n".join(f"{indent}{line}" if line else "" for line in lines)
-
-
-class _ExpressionParser:
-    def __init__(
-        self,
-        tokens: list[_Token],
-        enum_values: MutableMapping[str, dict[str, int]] | None = None,
-        enum_names: Iterable[str] | None = None,
-        scope_context: _ScopeContext | None = None,
-        macro_values: Mapping[str, str] | None = None,
-        macro_expansion_stack: frozenset[str] | None = None,
-    ) -> None:
-        self.tokens = tokens
-        self.position = 0
-        self.enum_values: MutableMapping[str, dict[str, int]] = (
-            enum_values if enum_values is not None else {}
-        )
-        self.enum_names: set[str] = set(enum_names or [])
-        self.scope_context = _normalize_scope_context(scope_context)
-        self.macro_values: MutableMapping[str, str] = dict(macro_values or {})
-        self.macro_expansion_stack = macro_expansion_stack or frozenset()
-
-    def parse(self) -> _Expression:
-        expr = self._parse_expression()
-        if not self._at_end():
-            raise GMLTranspileError(f"Unexpected token: {self._peek().value}")
-        return expr
-
-    def _parse_expression(self, min_precedence: int = 0) -> _Expression:
-        left = self._parse_unary()
-
-        while True:
-            if self._match("?"):
-                if _TERNARY_PRECEDENCE < min_precedence:
-                    self.position -= 1
-                    break
-                true_expr = self._parse_expression()
-                self._consume(":")
-                false_expr = self._parse_expression(_TERNARY_PRECEDENCE)
-                left = _Ternary(left, true_expr, false_expr)
-                continue
-
-            operator = self._current_operator()
-            if operator is None:
-                break
-
-            precedence = _BINARY_PRECEDENCE[operator]
-            if precedence < min_precedence:
-                break
-
-            self._advance()
-            next_precedence = precedence if operator in _RIGHT_ASSOCIATIVE else precedence + 1
-            right = self._parse_expression(next_precedence)
-            left = _Binary(left, operator, right)
-
-        return left
-
-    def _parse_unary(self) -> _Expression:
-        operator = self._current_unary_operator()
-        if operator in ("+", "-", "!", "not", "~"):
-            self._advance()
-            return _Unary(operator, self._parse_unary())
-        return self._parse_postfix()
-
-    def _parse_postfix(self) -> _Expression:
-        expr = self._parse_primary()
-        while True:
-            if self._match("("):
-                args: list[_Expression] = []
-                if not self._check(")"):
-                    while True:
-                        if self._check(",") or self._check(")"):
-                            args.append(_Name(_NAME_REPLACEMENTS["undefined"]))
-                        else:
-                            args.append(self._parse_expression())
-                        if not self._match(","):
-                            break
-                self._consume(")")
-                expr = _Call(expr, tuple(args))
-                continue
-
-            if self._match("["):
-                if self._match("$"):
-                    key = self._parse_expression()
-                    self._consume("]")
-                    expr = _StructAccess(expr, key)
-                    continue
-                if self._match("?"):
-                    key = self._parse_expression()
-                    self._consume("]")
-                    expr = _DSMapAccess(expr, key)
-                    continue
-                index = self._parse_expression()
-                self._consume("]")
-                expr = _Index(expr, index)
-                continue
-
-            if self._match("."):
-                member = self._consume_identifier()
-                expr = _Member(expr, member)
-                continue
-
-            break
-
-        return expr
-
-    def _parse_primary(self) -> _Expression:
-        token = self._advance()
-        if token.kind == "NUMBER":
-            return _NumberLiteral(token.value, _is_float_like_number(token.value))
-        if token.kind == "STRING":
-            return _StringLiteral(token.value)
-        if token.kind == "IDENT" and token.value == "nameof":
-            return self._parse_nameof_expression()
-        if token.kind == "IDENT" and token.value == "new":
-            return self._parse_new_expression()
-        if token.kind == "IDENT" and token.value == "function":
-            return self._parse_function_literal()
-        if token.kind == "IDENT":
-            macro_expr = self._parse_macro_expansion(token.value)
-            if macro_expr is not None:
-                return macro_expr
-            return _Name(_NAME_REPLACEMENTS.get(token.value, token.value))
-        if token.value == "[":
-            elements: list[_Expression] = []
-            if not self._check("]"):
-                while True:
-                    elements.append(self._parse_expression())
-                    if not self._match(","):
-                        break
-                    if self._check("]"):
-                        break
-            self._consume("]")
-            return _ArrayLiteral(tuple(elements))
-        if token.value == "{":
-            fields: list[tuple[str, _Expression]] = []
-            if not self._check("}"):
-                while True:
-                    field_name = self._consume_identifier()
-                    if self._match(":"):
-                        field_value = self._parse_expression()
-                    else:
-                        field_value = _Name(_NAME_REPLACEMENTS.get(field_name, field_name))
-                    fields.append((field_name, field_value))
-                    if not self._match(","):
-                        break
-                    if self._check("}"):
-                        break
-            self._consume("}")
-            return _StructLiteral(tuple(fields))
-        if token.value == "(":
-            expr = self._parse_expression()
-            self._consume(")")
-            return _Grouped(expr)
-        raise GMLTranspileError(f"Expected expression, got: {token.value}")
-
-    def _parse_macro_expansion(self, name: str) -> _Expression | None:
-        macro_source = self.macro_values.get(name)
-        if macro_source is None:
-            return None
-        if name in self.macro_expansion_stack:
-            raise GMLTranspileError(f"Recursive macro expansion for {name}")
-        return _parse_gml_expression(
-            macro_source,
-            self.enum_values,
-            self.enum_names,
-            self.macro_values,
-            self.macro_expansion_stack | {name},
-            scope_context=self.scope_context,
-        )
-
-    def _parse_new_expression(self) -> _Expression:
-        constructor = self._parse_primary()
-        while self._match("."):
-            constructor = _Member(constructor, self._consume_identifier())
-
-        self._consume("(")
-        args: list[_Expression] = []
-        if not self._check(")"):
-            while True:
-                if self._check(",") or self._check(")"):
-                    args.append(_Name(_NAME_REPLACEMENTS["undefined"]))
-                else:
-                    args.append(self._parse_expression())
-                if not self._match(","):
-                    break
-        self._consume(")")
-        return _NewCall(constructor, tuple(args))
-
-    def _parse_nameof_expression(self) -> _Expression:
-        self._consume("(")
-        name = self._parse_nameof_argument_name()
-        self._consume(")")
-        return _NameOf(name)
-
-    def _parse_nameof_argument_name(self) -> str:
-        token = self._advance()
-        if token.kind != "IDENT":
-            raise GMLTranspileError("nameof requires an identifier")
-
-        name = token.value
-        while self._match("."):
-            name = self._consume_identifier()
-
-        if self._match("("):
-            self._read_balanced_tokens("(", ")")
-
-        if not self._check(")"):
-            raise GMLTranspileError("nameof requires an identifier, enum member, or function-call syntax")
-        return name
-
-    def _parse_function_literal(self) -> _Expression:
-        name = None
-        if self._peek().kind == "IDENT":
-            name = self._consume_identifier()
-
-        parameters: list[_FunctionParameter] = []
-        self._consume("(")
-        if not self._check(")"):
-            while True:
-                parameter_name = self._consume_identifier()
-                default = self._parse_expression() if self._match("=") else None
-                parameters.append(_FunctionParameter(parameter_name, default))
-                if not self._match(","):
-                    break
-        self._consume(")")
-        parent_constructor: _Expression | None = None
-        if self._match(":"):
-            parent_constructor = self._parse_expression()
-        is_constructor = self._match_identifier("constructor")
-        self._consume("{")
-        body_tokens = self._read_balanced_tokens("{", "}")
-        static_declarations = _collect_static_declarations(body_tokens)
-        parameter_names = [parameter.name for parameter in parameters]
-        for parameter_name in parameter_names:
-            _validate_gml_identifier(parameter_name)
-            _reject_asset_identifier_name(parameter_name, self.scope_context)
-        scope_context = self.scope_context
-        static_scope_id = None
-        static_scope_name = None
-        static_prefix = scope_context.static_prefix
-        if static_declarations:
-            static_scope_id = _static_scope_id(static_prefix, name, self.position, body_tokens)
-            static_scope_name = (
-                f"_gml_static_scope_{hashlib.sha1(static_scope_id.encode('utf-8')).hexdigest()[:12]}"
-            )
-            static_prefix = static_scope_id
-        static_names = frozenset(declaration.name for declaration in static_declarations)
-        if is_constructor:
-            scope_context = _ScopeContext(
-                self_expression="_gml_constructor_self",
-                other_expression=self.scope_context.other_expression,
-                instance_target="_gml_constructor_self",
-                global_names=self.scope_context.global_names,
-                asset_names=self.scope_context.asset_names,
-                static_scope=static_scope_name,
-                static_names=static_names,
-                static_prefix=static_prefix,
-            )
-        else:
-            scope_context = _ScopeContext(
-                self_expression=scope_context.self_expression,
-                other_expression=scope_context.other_expression,
-                instance_target=scope_context.instance_target,
-                global_names=scope_context.global_names,
-                asset_names=scope_context.asset_names,
-                static_scope=static_scope_name,
-                static_names=static_names,
-                static_prefix=static_prefix,
-            )
-        body_parser = _StatementParser(
-            body_tokens,
-            local_names=parameter_names,
-            return_depth=1,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=scope_context,
-            macro_values=self.macro_values,
-            global_names=scope_context.global_names,
-        )
-        body_lines = body_parser.parse()
-        prelude_lines: list[str] = []
-        if parent_constructor is not None:
-            if not is_constructor:
-                raise GMLTranspileError("Constructor inheritance requires a constructor function")
-            prelude_lines.append(
-                _emit_constructor_inheritance_line(
-                    parent_constructor,
-                    parameter_names,
-                    scope_context,
-                    self.scope_context,
-                )
-            )
-        if static_declarations:
-            prelude_lines.extend(
-                _emit_static_initialization_lines(
-                    static_scope_name,
-                    static_scope_id,
-                    static_declarations,
-                    parameter_names,
-                    scope_context,
-                    self.enum_values,
-                    self.enum_names,
-                    self.macro_values,
-                )
-            )
-        if prelude_lines:
-            body_lines = [*prelude_lines, *body_lines]
-        return _FunctionLiteral(
-            name,
-            tuple(parameters),
-            tuple(body_lines or ["pass"]),
-            is_constructor,
-            static_scope_id if static_declarations else None,
-        )
-
-    def _current_operator(self) -> str | None:
-        token = self._peek()
-        if token.kind == "IDENT" and token.value in _BINARY_PRECEDENCE:
-            return token.value
-        if token.value in _BINARY_PRECEDENCE:
-            return token.value
-        return None
-
-    def _current_unary_operator(self) -> str | None:
-        token = self._peek()
-        if token.kind == "IDENT" and token.value == "not":
-            return token.value
-        if token.value in ("+", "-", "!", "~"):
-            return token.value
-        return None
-
-    def _match(self, value: str) -> bool:
-        if self._check(value):
-            self.position += 1
-            return True
-        return False
-
-    def _match_identifier(self, value: str) -> bool:
-        token = self._peek()
-        if token.kind == "IDENT" and token.value == value:
-            self.position += 1
-            return True
-        return False
-
-    def _consume(self, value: str) -> None:
-        if not self._match(value):
-            raise GMLTranspileError(f"Expected '{value}', got: {self._peek().value}")
-
-    def _consume_identifier(self) -> str:
-        token = self._advance()
-        if token.kind != "IDENT":
-            raise GMLTranspileError(f"Expected identifier, got: {token.value}")
-        _validate_gml_identifier(token.value)
-        return token.value
-
-    def _read_balanced_tokens(self, opener: str, closer: str) -> list[_Token]:
-        tokens: list[_Token] = []
-        depth = 1
-        while not self._at_end():
-            token = self._advance()
-            if token.value == opener:
-                depth += 1
-            elif token.value == closer:
-                depth -= 1
-                if depth == 0:
-                    return tokens
-            tokens.append(token)
-
-        raise GMLTranspileError(f"Expected '{closer}'")
-
-    def _check(self, value: str) -> bool:
-        return self._peek().value == value
-
-    def _advance(self) -> _Token:
-        token = self._peek()
-        if not self._at_end():
-            self.position += 1
-        return token
-
-    def _peek(self) -> _Token:
-        if self.position >= len(self.tokens):
-            return _EOF
-        return self.tokens[self.position]
-
-    def _at_end(self) -> bool:
-        return self._peek().kind == "EOF"
-
-
-class _StatementParser:
-    def __init__(
-        self,
-        tokens: list[_Token],
-        local_names: Iterable[str] | None = None,
-        instance_variables: MutableSet[str] | None = None,
-        loop_depth: int = 0,
-        continue_depth: int = 0,
-        return_depth: int = 0,
-        finally_depth: int = 0,
-        generated_counter: list[int] | None = None,
-        enum_values: MutableMapping[str, dict[str, int]] | None = None,
-        enum_names: Iterable[str] | None = None,
-        scope_context: _ScopeContext | None = None,
-        inherited_event_call: str | None = None,
-        macro_values: MutableMapping[str, str] | None = None,
-        macro_priorities: MutableMapping[str, int] | None = None,
-        macro_configuration: str | None = None,
-        top_level_global_scope: bool = False,
-        global_names: Iterable[str] | None = None,
-        asset_names: Iterable[str] | None = None,
-        static_scope_prefix: str | None = None,
-    ) -> None:
-        self.tokens = tokens
-        self.position = 0
-        self.local_names = set(local_names or [])
-        self.instance_variables = instance_variables
-        self.loop_depth = loop_depth
-        self.continue_depth = continue_depth
-        self.return_depth = return_depth
-        self.finally_depth = finally_depth
-        self.generated_counter = generated_counter if generated_counter is not None else [0]
-        self.enum_values: MutableMapping[str, dict[str, int]] = (
-            enum_values if enum_values is not None else {}
-        )
-        self.enum_names: set[str] = set(enum_names or [])
-        self.global_names: set[str] = set(global_names or [])
-        self.scope_context = _scope_context_with_global_names(
-            _normalize_scope_context(scope_context),
-            self.global_names,
-            top_level_global_scope=top_level_global_scope,
-            asset_names=asset_names,
-            static_prefix=static_scope_prefix,
-        )
-        self.inherited_event_call = inherited_event_call
-        self.macro_values: MutableMapping[str, str] = (
-            macro_values if macro_values is not None else {}
-        )
-        self.macro_priorities: MutableMapping[str, int] = (
-            macro_priorities if macro_priorities is not None else {}
-        )
-        self.macro_configuration = macro_configuration
-
-    def parse(self, terminator: str | None = None) -> list[str]:
-        lines: list[str] = []
-        while not self._at_end() and not self._check(terminator):
-            if self._match(";") or self._match("\n"):
-                continue
-            lines.extend(self._parse_statement())
-        return lines
-
-    def _parse_statement(self) -> list[str]:
-        if self._check_directive("#macro"):
-            return self._parse_macro_statement()
-        if self._check_identifier("globalvar"):
-            return self._parse_globalvar_statement()
-        if self._check_identifier("static"):
-            return self._parse_static_statement()
-        if self._check_identifier("enum"):
-            return self._parse_enum_statement()
-        if self._check_identifier("if"):
-            return self._parse_if_statement()
-        if self._check_identifier("with"):
-            return self._parse_with_statement()
-        if self._check_identifier("while"):
-            return self._parse_while_statement()
-        if self._check_identifier("repeat"):
-            return self._parse_repeat_statement()
-        if self._check_identifier("do"):
-            return self._parse_do_until_statement()
-        if self._check_identifier("for"):
-            return self._parse_for_statement()
-        if self._check_identifier("switch"):
-            return self._parse_switch_statement()
-        if self._check_identifier("try"):
-            return self._parse_try_statement()
-
-        if self._match("{"):
-            lines = self.parse(terminator="}")
-            self._consume("}")
-            return lines
-
-        statement_tokens = self._read_simple_statement()
-        if not statement_tokens:
-            return []
-        return _transpile_statement(
-            _tokens_to_source(statement_tokens),
-            self.local_names,
-            self.instance_variables,
-            loop_depth=self.loop_depth,
-            continue_depth=self.continue_depth,
-            return_depth=self.return_depth,
-            finally_depth=self.finally_depth,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=self.scope_context,
-            inherited_event_call=self.inherited_event_call,
-            macro_values=self.macro_values,
-            generated_counter=self.generated_counter,
-        )
-
-    def _parse_macro_statement(self) -> list[str]:
-        self._consume_directive("#macro")
-        configuration_or_name = self._consume_identifier_name()
-        configuration: str | None = None
-        name = configuration_or_name
-        if self._match(":"):
-            configuration = configuration_or_name
-            name = self._consume_identifier_name()
-
-        value_tokens = self._read_macro_value_tokens()
-        value_source = _tokens_to_source(value_tokens)
-        if not value_source:
-            raise GMLTranspileError("Expected macro value")
-
-        priority = 0
-        if configuration is not None:
-            if not _macro_configuration_matches(configuration, self.macro_configuration):
-                return []
-            priority = 1
-
-        current_priority = self.macro_priorities.get(name, -1)
-        if priority >= current_priority:
-            self.macro_values[name] = value_source
-            self.macro_priorities[name] = priority
-        return []
-
-    def _parse_globalvar_statement(self) -> list[str]:
-        self._consume_identifier("globalvar")
-        while not self._at_end() and not self._check(";") and not self._check("\n"):
-            name = self._consume_identifier_name()
-            _validate_gml_identifier(name)
-            _reject_asset_identifier_name(name, self.scope_context)
-            self.global_names.add(name)
-            if not self._match(","):
-                break
-        self._match(";")
-        self._match("\n")
-        self.scope_context = _scope_context_with_global_names(
-            self.scope_context,
-            self.global_names,
-        )
-        return []
-
-    def _parse_static_statement(self) -> list[str]:
-        self._read_simple_statement()
-        if self.scope_context.static_scope is None:
-            raise GMLTranspileError("static declarations are only supported inside functions")
-        return []
-
-    def _parse_enum_statement(self) -> list[str]:
-        self._consume_identifier("enum")
-        enum_name = self._consume_identifier_name()
-        gdscript_enum_name = _sanitize_gdscript_identifier(enum_name)
-        self._skip_newlines()
-        self._consume("{")
-
-        members: list[tuple[str, str]] = []
-        current_enum_values: dict[str, int] = {}
-        next_integer_value = 0
-        while not self._at_end() and not self._check("}"):
-            if self._match(",") or self._match(";") or self._match("\n"):
-                continue
-            member_name = self._consume_identifier_name()
-            if self._match("="):
-                value_tokens = self._read_enum_value_tokens()
-                enum_value = _evaluate_enum_value_tokens(
-                    value_tokens,
-                    self.enum_values,
-                    current_enum_values,
-                    self.macro_values,
-                )
-                next_integer_value = enum_value + 1
-            else:
-                enum_value = next_integer_value
-                next_integer_value += 1
-            value = str(enum_value)
-            current_enum_values[member_name] = enum_value
-            members.append((member_name, value))
-
-        self._consume("}")
-        self.enum_values[enum_name] = current_enum_values
-        self.enum_names.add(enum_name)
-        self.local_names.add(enum_name)
-
-        fields = ", ".join(f"{json.dumps(member_name)}: {value}" for member_name, value in members)
-        return [f"var {gdscript_enum_name} = GMRuntime.gml_enum({{{fields}}})"]
-
-    def _parse_if_statement(self) -> list[str]:
-        self._consume_identifier("if")
-        condition_tokens = self._read_condition_tokens()
-        if not condition_tokens:
-            raise GMLTranspileError("Expected if condition")
-
-        condition = transpile_gml_condition(
-            _tokens_to_source(condition_tokens),
-            local_names=self.local_names,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=self.scope_context,
-            macro_values=self.macro_values,
-        )
-        body_lines = self._parse_body()
-        lines = [f"if {condition}:"]
-        lines.extend(_indent_lines(body_lines or ["pass"]))
-
-        self._skip_newlines()
-        if self._match_identifier("else"):
-            if self._check_identifier("if"):
-                else_lines = self._parse_if_statement()
-                lines.append(f"elif {else_lines[0][3:]}")
-                lines.extend(else_lines[1:])
-            else:
-                else_body_lines = self._parse_body()
-                lines.append("else:")
-                lines.extend(_indent_lines(else_body_lines or ["pass"]))
-
-        return lines
-
-    def _parse_with_statement(self) -> list[str]:
-        self._consume_identifier("with")
-        target_tokens = self._read_condition_tokens()
-        if not target_tokens:
-            raise GMLTranspileError("Expected with target")
-
-        target = transpile_gml_expression(
-            _tokens_to_source(target_tokens),
-            local_names=self.local_names,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=self.scope_context,
-            macro_values=self.macro_values,
-        )
-        with_target = self._next_generated_name("_gml_with_target")
-        outer_scope_context = self.scope_context
-        self.scope_context = _ScopeContext(
-            self_expression=with_target,
-            other_expression=outer_scope_context.self_expression,
-            instance_target=with_target,
-            global_names=outer_scope_context.global_names,
-            asset_names=outer_scope_context.asset_names,
-            static_scope=outer_scope_context.static_scope,
-            static_names=outer_scope_context.static_names,
-            static_prefix=outer_scope_context.static_prefix,
-        )
-        self.loop_depth += 1
-        self.continue_depth += 1
-        try:
-            body_lines = self._parse_body()
-        finally:
-            self.loop_depth -= 1
-            self.continue_depth -= 1
-            self.scope_context = outer_scope_context
-
-        lines = [
-            "for "
-            f"{with_target} in "
-            f"GMRuntime.gml_with_targets({target}, "
-            f"{outer_scope_context.self_expression}, {outer_scope_context.other_expression}):"
-        ]
-        lines.extend(_indent_lines(body_lines or ["pass"]))
-        return lines
-
-    def _parse_while_statement(self) -> list[str]:
-        self._consume_identifier("while")
-        condition_tokens = self._read_condition_tokens()
-        if not condition_tokens:
-            raise GMLTranspileError("Expected while condition")
-
-        condition = transpile_gml_condition(
-            _tokens_to_source(condition_tokens),
-            local_names=self.local_names,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=self.scope_context,
-            macro_values=self.macro_values,
-        )
-        self.loop_depth += 1
-        self.continue_depth += 1
-        try:
-            body_lines = self._parse_body()
-        finally:
-            self.loop_depth -= 1
-            self.continue_depth -= 1
-        lines = [f"while {condition}:"]
-        lines.extend(_indent_lines(body_lines or ["pass"]))
-        return lines
-
-    def _parse_repeat_statement(self) -> list[str]:
-        self._consume_identifier("repeat")
-        count_tokens = self._read_condition_tokens()
-        if not count_tokens:
-            raise GMLTranspileError("Expected repeat count")
-
-        count = transpile_gml_expression(
-            _tokens_to_source(count_tokens),
-            local_names=self.local_names,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=self.scope_context,
-            macro_values=self.macro_values,
-        )
-        self.loop_depth += 1
-        self.continue_depth += 1
-        try:
-            body_lines = self._parse_body()
-        finally:
-            self.loop_depth -= 1
-            self.continue_depth -= 1
-
-        lines = [f"for _gml_repeat_index in range(GMRuntime.gml_repeat_count({count})):"]
-        lines.extend(_indent_lines(body_lines or ["pass"]))
-        return lines
-
-    def _parse_for_statement(self) -> list[str]:
-        self._consume_identifier("for")
-        self._consume("(")
-        header_tokens = self._read_balanced_tokens("(", ")")
-        header_parts = _split_top_level_tokens(header_tokens, ";")
-        if len(header_parts) != 3:
-            raise GMLTranspileError("Expected for initializer, condition, and operation clauses")
-
-        initializer = _tokens_to_source(header_parts[0])
-        condition_source = _tokens_to_source(header_parts[1])
-        operation = _tokens_to_source(header_parts[2])
-
-        lines: list[str] = []
-        if initializer:
-            lines.extend(
-                _transpile_statement(
-                    initializer,
-                    self.local_names,
-                    self.instance_variables,
-                    loop_depth=self.loop_depth,
-                    continue_depth=self.continue_depth,
-                    return_depth=self.return_depth,
-                    finally_depth=self.finally_depth,
-                    enum_values=self.enum_values,
-                    enum_names=self.enum_names,
-                    scope_context=self.scope_context,
-                    inherited_event_call=self.inherited_event_call,
-                    macro_values=self.macro_values,
-                    generated_counter=self.generated_counter,
-                )
-            )
-
-        condition = (
-            transpile_gml_condition(
-                condition_source,
-                local_names=self.local_names,
-                enum_values=self.enum_values,
-                enum_names=self.enum_names,
-                scope_context=self.scope_context,
-                macro_values=self.macro_values,
-            )
-            if condition_source
-            else "true"
-        )
-        operation_lines = (
-            _transpile_statement(
-                operation,
-                self.local_names,
-                self.instance_variables,
-                loop_depth=self.loop_depth,
-                continue_depth=self.continue_depth,
-                return_depth=self.return_depth,
-                finally_depth=self.finally_depth,
-                enum_values=self.enum_values,
-                enum_names=self.enum_names,
-                scope_context=self.scope_context,
-                inherited_event_call=self.inherited_event_call,
-                macro_values=self.macro_values,
-                generated_counter=self.generated_counter,
-            )
-            if operation
-            else []
-        )
-
-        self.loop_depth += 1
-        self.continue_depth += 1
-        try:
-            body_lines = self._parse_body()
-        finally:
-            self.loop_depth -= 1
-            self.continue_depth -= 1
-
-        if operation_lines:
-            body_lines = _insert_lines_before_continue(body_lines, operation_lines)
-            body_lines.extend(operation_lines)
-
-        lines.append(f"while {condition}:")
-        lines.extend(_indent_lines(body_lines or ["pass"]))
-        return lines
-
-    def _parse_do_until_statement(self) -> list[str]:
-        self._consume_identifier("do")
-        self.loop_depth += 1
-        self.continue_depth += 1
-        try:
-            body_lines = self._parse_do_until_body()
-        finally:
-            self.loop_depth -= 1
-            self.continue_depth -= 1
-
-        self._skip_newlines()
-        self._consume_identifier("until")
-        condition_tokens = self._read_condition_tokens()
-        if not condition_tokens:
-            raise GMLTranspileError("Expected until condition")
-
-        condition = transpile_gml_condition(
-            _tokens_to_source(condition_tokens),
-            local_names=self.local_names,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=self.scope_context,
-            macro_values=self.macro_values,
-        )
-        body_lines = _insert_until_check_before_continue(body_lines, condition)
-        lines = ["while true:"]
-        lines.extend(_indent_lines(body_lines or ["pass"]))
-        lines.append(f"\tif {condition}:")
-        lines.append("\t\tbreak")
-        return lines
-
-    def _parse_switch_statement(self) -> list[str]:
-        self._consume_identifier("switch")
-        expression_tokens = self._read_condition_tokens()
-        if not expression_tokens:
-            raise GMLTranspileError("Expected switch expression")
-
-        switch_value = self._next_generated_name("_gml_switch_value")
-        switch_matched = self._next_generated_name("_gml_switch_matched")
-        switch_has_case = self._next_generated_name("_gml_switch_has_case")
-        expression = transpile_gml_expression(
-            _tokens_to_source(expression_tokens),
-            local_names=self.local_names,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=self.scope_context,
-            macro_values=self.macro_values,
-        )
-        sections = self._parse_switch_sections()
-        case_values = [
-            section_value
-            for section_kind, section_value, _ in sections
-            if section_kind == "case" and section_value is not None
-        ]
-        case_conditions = [
-            f"GMRuntime.gml_eq({switch_value}, {case_value})"
-            for case_value in case_values
-        ]
-        has_case_expression = " or ".join(case_conditions) if case_conditions else "false"
-
-        lines = [
-            f"var {switch_value} = {expression}",
-            f"var {switch_matched} = false",
-            f"var {switch_has_case} = {has_case_expression}",
-            "while true:",
-        ]
-        for section_kind, section_value, section_lines in sections:
-            if section_kind == "case":
-                lines.append(
-                    f"\tif not {switch_matched} and GMRuntime.gml_eq({switch_value}, {section_value}):"
-                )
-                lines.append(f"\t\t{switch_matched} = true")
-            else:
-                lines.append(f"\tif not {switch_matched} and not {switch_has_case}:")
-                lines.append(f"\t\t{switch_matched} = true")
-
-            lines.append(f"\tif {switch_matched}:")
-            lines.extend(f"\t\t{line}" for line in (section_lines or ["pass"]))
-        lines.append("\tbreak")
-        return lines
-
-    def _parse_switch_sections(self) -> list[tuple[str, str | None, list[str]]]:
-        self._skip_newlines()
-        self._consume("{")
-        sections: list[tuple[str, str | None, list[str]]] = []
-        while not self._at_end() and not self._check("}"):
-            if self._match(";") or self._match("\n"):
-                continue
-            if self._match_identifier("case"):
-                label_tokens = self._read_switch_label_tokens()
-                if not label_tokens:
-                    raise GMLTranspileError("Expected switch case value")
-                label = transpile_gml_expression(
-                    _tokens_to_source(label_tokens),
-                    local_names=self.local_names,
-                    enum_values=self.enum_values,
-                    enum_names=self.enum_names,
-                    scope_context=self.scope_context,
-                    macro_values=self.macro_values,
-                )
-                body_tokens = self._read_switch_section_body_tokens()
-                sections.append(("case", label, self._parse_switch_section_body(body_tokens)))
-                continue
-            if self._match_identifier("default"):
-                self._consume(":")
-                body_tokens = self._read_switch_section_body_tokens()
-                sections.append(("default", None, self._parse_switch_section_body(body_tokens)))
-                continue
-            raise GMLTranspileError(f"Expected switch case or default, got: {self._peek().value}")
-
-        self._consume("}")
-        return sections
-
-    def _parse_switch_section_body(self, body_tokens: list[_Token]) -> list[str]:
-        parser = _StatementParser(
-            body_tokens,
-            local_names=self.local_names,
-            instance_variables=self.instance_variables,
-            loop_depth=self.loop_depth + 1,
-            continue_depth=self.continue_depth,
-            return_depth=self.return_depth,
-            finally_depth=self.finally_depth,
-            generated_counter=self.generated_counter,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=self.scope_context,
-            inherited_event_call=self.inherited_event_call,
-            macro_values=self.macro_values,
-            macro_priorities=self.macro_priorities,
-            macro_configuration=self.macro_configuration,
-            global_names=self.global_names,
-        )
-        lines = parser.parse()
-        self.local_names.update(parser.local_names)
-        self.enum_names.update(parser.enum_names)
-        self.global_names.update(parser.global_names)
-        self.scope_context = _scope_context_with_global_names(
-            self.scope_context,
-            self.global_names,
-        )
-        return lines
-
-    def _parse_try_statement(self) -> list[str]:
-        self._consume_identifier("try")
-        exception_name = self._next_generated_name("_gml_try_exception")
-        try_lines = self._parse_body()
-
-        self._skip_newlines()
-        catch_name: str | None = None
-        catch_lines: list[str] | None = None
-        if self._match_identifier("catch"):
-            catch_name = self._parse_catch_variable_name()
-            had_catch_local = catch_name in self.local_names
-            self.local_names.add(catch_name)
-            try:
-                catch_lines = self._parse_body()
-            finally:
-                if not had_catch_local:
-                    self.local_names.discard(catch_name)
-
-        self._skip_newlines()
-        finally_lines: list[str] | None = None
-        if self._match_identifier("finally"):
-            self.finally_depth += 1
-            try:
-                finally_lines = self._parse_body()
-            finally:
-                self.finally_depth -= 1
-
-        if catch_lines is None and finally_lines is None:
-            raise GMLTranspileError("try requires catch or finally")
-
-        lines = [
-            f"var {exception_name} = (func():",
-            *_indent_lines(try_lines or ["pass"]),
-            "\treturn GMRuntime.gml_undefined()",
-            ").call()",
-        ]
-
-        if catch_lines is not None and catch_name is not None:
-            gdscript_catch_name = _sanitize_gdscript_identifier(catch_name)
-            lines.extend([
-                f"if GMRuntime.is_gml_exception({exception_name}):",
-                f"\tvar {gdscript_catch_name} = GMRuntime.gml_exception_struct({exception_name})",
-                f"\t{exception_name} = (func():",
-                *[f"\t\t{line}" if line else "" for line in (catch_lines or ["pass"])],
-                "\t\treturn GMRuntime.gml_undefined()",
-                "\t).call()",
-            ])
-
-        if finally_lines is not None:
-            lines.extend(finally_lines or ["pass"])
-
-        lines.extend([
-            f"if GMRuntime.is_gml_exception({exception_name}):",
-            f"\treturn {exception_name}",
-        ])
-        return lines
-
-    def _parse_catch_variable_name(self) -> str:
-        catch_tokens = self._read_condition_tokens()
-        if len(catch_tokens) != 1 or catch_tokens[0].kind != "IDENT":
-            raise GMLTranspileError("catch requires a variable name")
-        catch_name = catch_tokens[0].value
-        _validate_gml_identifier(catch_name)
-        return catch_name
-
-    def _read_switch_label_tokens(self) -> list[_Token]:
-        tokens: list[_Token] = []
-        depth = 0
-        while not self._at_end():
-            token = self._peek()
-            if depth == 0 and token.value == ":":
-                self._advance()
-                return tokens
-
-            if token.value in "([{":
-                depth += 1
-            elif token.value in ")]}" and depth > 0:
-                depth -= 1
-            tokens.append(self._advance())
-
-        raise GMLTranspileError("Expected ':' after switch case")
-
-    def _read_switch_section_body_tokens(self) -> list[_Token]:
-        tokens: list[_Token] = []
-        depth = 0
-        while not self._at_end():
-            token = self._peek()
-            if depth == 0 and token.value == "}":
-                break
-            if depth == 0 and token.kind == "IDENT" and token.value in ("case", "default"):
-                break
-
-            if token.value in "([{":
-                depth += 1
-            elif token.value in ")]}":
-                if depth > 0:
-                    depth -= 1
-            tokens.append(self._advance())
-
-        return tokens
-
-    def _parse_do_until_body(self) -> list[str]:
-        if self._match("{"):
-            lines = self.parse(terminator="}")
-            self._consume("}")
-            return lines
-
-        statement_tokens = self._read_do_until_statement_tokens()
-        if not statement_tokens:
-            return []
-        return _transpile_statement(
-            _tokens_to_source(statement_tokens),
-            self.local_names,
-            self.instance_variables,
-            loop_depth=self.loop_depth,
-            continue_depth=self.continue_depth,
-            return_depth=self.return_depth,
-            finally_depth=self.finally_depth,
-            enum_values=self.enum_values,
-            enum_names=self.enum_names,
-            scope_context=self.scope_context,
-            inherited_event_call=self.inherited_event_call,
-            macro_values=self.macro_values,
-            generated_counter=self.generated_counter,
-        )
-
-    def _parse_body(self) -> list[str]:
-        if self._match("{"):
-            lines = self.parse(terminator="}")
-            self._consume("}")
-            return lines
-        if self._at_end() or self._check("}"):
-            return []
-        return self._parse_statement()
-
-    def _read_condition_tokens(self) -> list[_Token]:
-        if self._match("("):
-            return self._read_balanced_tokens("(", ")")
-
-        tokens: list[_Token] = []
-        depth = 0
-        while not self._at_end():
-            token = self._peek()
-            if depth == 0 and token.value == "{":
-                break
-            if depth == 0 and token.value == ";":
-                break
-
-            if token.value in "([":
-                depth += 1
-            elif token.value in ")]" and depth > 0:
-                depth -= 1
-            tokens.append(self._advance())
-
-        return tokens
-
-    def _read_do_until_statement_tokens(self) -> list[_Token]:
-        tokens: list[_Token] = []
-        depth = 0
-        while not self._at_end():
-            token = self._peek()
-            if depth == 0 and token.kind == "IDENT" and token.value == "until":
-                break
-
-            if token.value in "([":
-                depth += 1
-            elif token.value in ")]" and depth > 0:
-                depth -= 1
-            tokens.append(self._advance())
-
-        return tokens
-
-    def _read_enum_value_tokens(self) -> list[_Token]:
-        tokens: list[_Token] = []
-        depth = 0
-        while not self._at_end():
-            token = self._peek()
-            if depth == 0 and token.value in (",", "}", "\n"):
-                break
-            if token.value in "([":
-                depth += 1
-            elif token.value in ")]" and depth > 0:
-                depth -= 1
-            tokens.append(self._advance())
-        if not tokens:
-            raise GMLTranspileError("Expected enum value")
-        return tokens
-
-    def _read_macro_value_tokens(self) -> list[_Token]:
-        tokens: list[_Token] = []
-        while not self._at_end() and not self._check("\n"):
-            tokens.append(self._advance())
-        self._match("\n")
-        return tokens
-
-    def _skip_newlines(self) -> None:
-        while self._match("\n"):
-            pass
-
-    def _read_balanced_tokens(self, opener: str, closer: str) -> list[_Token]:
-        tokens: list[_Token] = []
-        depth = 1
-        while not self._at_end():
-            token = self._advance()
-            if token.value == opener:
-                depth += 1
-            elif token.value == closer:
-                depth -= 1
-                if depth == 0:
-                    return tokens
-            tokens.append(token)
-
-        raise GMLTranspileError(f"Expected '{closer}'")
-
-    def _read_simple_statement(self) -> list[_Token]:
-        tokens: list[_Token] = []
-        depth = 0
-        while not self._at_end():
-            token = self._peek()
-            if depth == 0 and token.value == "}":
-                break
-            if depth == 0 and token.value in (";", "\n"):
-                self._advance()
-                break
-
-            if token.value in "([{":
-                depth += 1
-            elif token.value in ")]}":
-                if depth > 0:
-                    depth -= 1
-            tokens.append(self._advance())
-
-        return tokens
-
-    def _match(self, value: str) -> bool:
-        if self._check(value):
-            self.position += 1
-            return True
-        return False
-
-    def _match_identifier(self, value: str) -> bool:
-        if self._check_identifier(value):
-            self.position += 1
-            return True
-        return False
-
-    def _consume(self, value: str) -> None:
-        if not self._match(value):
-            raise GMLTranspileError(f"Expected '{value}', got: {self._peek().value}")
-
-    def _consume_identifier(self, value: str) -> None:
-        if not self._match_identifier(value):
-            raise GMLTranspileError(f"Expected '{value}', got: {self._peek().value}")
-
-    def _consume_directive(self, value: str) -> None:
-        if not self._check_directive(value):
-            raise GMLTranspileError(f"Expected '{value}', got: {self._peek().value}")
-        self.position += 1
-
-    def _consume_identifier_name(self) -> str:
-        token = self._advance()
-        if token.kind != "IDENT":
-            raise GMLTranspileError(f"Expected identifier, got: {token.value}")
-        _validate_gml_identifier(token.value)
-        return token.value
-
-    def _check(self, value: str | None) -> bool:
-        if value is None:
-            return False
-        return self._peek().value == value
-
-    def _check_identifier(self, value: str) -> bool:
-        token = self._peek()
-        return token.kind == "IDENT" and token.value == value
-
-    def _check_directive(self, value: str) -> bool:
-        token = self._peek()
-        return token.kind == "DIRECTIVE" and token.value == value
-
-    def _advance(self) -> _Token:
-        token = self._peek()
-        if not self._at_end():
-            self.position += 1
-        return token
-
-    def _peek(self) -> _Token:
-        if self.position >= len(self.tokens):
-            return _EOF
-        return self.tokens[self.position]
-
-    def _at_end(self) -> bool:
-        return self._peek().kind == "EOF"
-
-    def _next_generated_name(self, prefix: str) -> str:
-        index = self.generated_counter[0]
-        self.generated_counter[0] += 1
-        return f"{prefix}_{index}"
-
-
-def _tokenize(source: str) -> list[_Token]:
-    tokens: list[_Token] = []
-    index = 0
-    while index < len(source):
-        char = source[index]
-
-        if char in "\r\n":
-            if char == "\r" and index + 1 < len(source) and source[index + 1] == "\n":
-                index += 1
-            tokens.append(_Token("NEWLINE", "\n"))
-            index += 1
-            continue
-
-        if char.isspace():
-            index += 1
-            continue
-
-        if char.isdigit() or (char == "." and index + 1 < len(source) and source[index + 1].isdigit()):
-            number_end = _read_number(source, index)
-            tokens.append(_Token("NUMBER", source[index:number_end].replace("_", "")))
-            index = number_end
-            continue
-
-        if char == '"' or char == "'":
-            tokens.append(_Token("STRING", _read_string(source, index)))
-            index += len(tokens[-1].value)
-            continue
-
-        if char == "$":
-            next_char = source[index + 1] if index + 1 < len(source) else ""
-            if next_char.lower() in "0123456789abcdef" or next_char == "_":
-                hex_end = _read_hex_number(source, index + 1)
-                tokens.append(_Token("NUMBER", f"0x{source[index + 1:hex_end].replace('_', '')}"))
-                index = hex_end
-            else:
-                tokens.append(_Token("OP", char))
-                index += 1
-            continue
-
-        if char == "#":
-            if _source_startswith_directive(source, index, "#macro"):
-                tokens.append(_Token("DIRECTIVE", "#macro"))
-                index += len("#macro")
-                continue
-            color_literal, color_end = _read_hash_color_literal(source, index)
-            tokens.append(_Token("NUMBER", color_literal))
-            index = color_end
-            continue
-
-        if char.isalpha() or char == "_":
-            start = index
-            index += 1
-            while index < len(source) and (source[index].isalnum() or source[index] == "_"):
-                index += 1
-            identifier = source[start:index]
-            _validate_gml_identifier(identifier)
-            block_delimiter = _BLOCK_DELIMITER_REPLACEMENTS.get(identifier)
-            if block_delimiter is not None:
-                tokens.append(_Token("OP", block_delimiter))
-            else:
-                tokens.append(_Token("IDENT", identifier))
-            continue
-
-        matched_operator = None
-        for operator in _MULTI_CHAR_OPERATORS:
-            if source.startswith(operator, index):
-                matched_operator = operator
-                break
-        if matched_operator is not None:
-            tokens.append(_Token("OP", matched_operator))
-            index += len(matched_operator)
-            continue
-
-        if char in "+-*/%&|^~!=<>()[]{}?:,.;.":
-            tokens.append(_Token("OP", char))
-            index += 1
-            continue
-
-        raise GMLTranspileError(f"Unexpected character: {char}")
-
-    tokens.append(_EOF)
-    return tokens
-
-
-def _source_startswith_directive(source: str, index: int, directive: str) -> bool:
-    if not source.startswith(directive, index):
-        return False
-    end = index + len(directive)
-    if end >= len(source):
-        return True
-    next_char = source[end]
-    return not (next_char.isalnum() or next_char == "_")
-
-
-def _expression_tokens(source: str) -> list[_Token]:
-    return [token for token in _tokenize(source) if token.kind != "NEWLINE"]
-
-
-def _read_number(source: str, start: int) -> int:
-    if source.startswith(("0b", "0B"), start):
-        return _read_binary_number(source, start)
-
-    if source.startswith(("0x", "0X"), start):
-        return _read_hex_number(source, start + 2)
-
-    return _read_decimal_number(source, start)
-
-
-def _read_decimal_number(source: str, start: int) -> int:
-    index = start
-    if source[index] == ".":
-        index += 1
-        index, saw_fraction_digit = _read_separated_digits(
-            source,
-            index,
-            "0123456789",
-            "numeric",
-        )
-        if not saw_fraction_digit:
-            raise GMLTranspileError("Malformed numeric literal")
-        return index
-
-    index, _saw_integer_digit = _read_separated_digits(
-        source,
-        index,
-        "0123456789",
-        "numeric",
-    )
-    if index < len(source) and source[index] == ".":
-        index += 1
-        if index < len(source) and source[index] == "_":
-            raise GMLTranspileError("Malformed numeric literal")
-        index, _saw_fraction_digit = _read_separated_digits(
-            source,
-            index,
-            "0123456789",
-            "numeric",
-        )
-
-    return index
-
-
-def _read_hex_number(source: str, start: int) -> int:
-    index, saw_digit = _read_separated_digits(
-        source,
-        start,
-        "0123456789abcdef",
-        "hexadecimal",
-    )
-
-    if not saw_digit:
-        raise GMLTranspileError("Malformed hexadecimal literal")
-
-    if index < len(source) and source[index].isalnum():
-        raise GMLTranspileError(f"Invalid hexadecimal literal digit: {source[index]}")
-
-    return index
-
-
-def _read_hash_color_literal(source: str, start: int) -> tuple[str, int]:
-    hex_start = start + 1
-    hex_end = hex_start + 6
-    if hex_end > len(source):
-        raise GMLTranspileError("Malformed hash color literal")
-
-    value = source[hex_start:hex_end]
-    if any(char.lower() not in "0123456789abcdef" for char in value):
-        raise GMLTranspileError("Malformed hash color literal")
-
-    if hex_end < len(source) and source[hex_end].isalnum():
-        raise GMLTranspileError(f"Invalid hash color literal digit: {source[hex_end]}")
-
-    red = value[0:2]
-    green = value[2:4]
-    blue = value[4:6]
-    return f"0x{blue}{green}{red}".lower(), hex_end
-
-
-def _read_binary_number(source: str, start: int) -> int:
-    index, saw_digit = _read_separated_digits(
-        source,
-        start + 2,
-        "01",
-        "binary",
-    )
-
-    if not saw_digit:
-        raise GMLTranspileError("Malformed binary literal")
-
-    if index < len(source) and source[index].isalnum():
-        raise GMLTranspileError(f"Invalid binary literal digit: {source[index]}")
-
-    return index
-
-
-def _read_separated_digits(
-    source: str,
-    start: int,
-    valid_digits: str,
-    literal_name: str,
-) -> tuple[int, bool]:
-    index = start
-    saw_digit = False
-    previous_was_digit = False
-
-    while index < len(source):
-        char = source[index]
-        if char.lower() in valid_digits:
-            saw_digit = True
-            previous_was_digit = True
-            index += 1
-            continue
-
-        if char == "_":
-            next_char = source[index + 1] if index + 1 < len(source) else ""
-            if (
-                not previous_was_digit
-                or next_char == ""
-                or next_char.lower() not in valid_digits
-            ):
-                raise GMLTranspileError(f"Malformed {literal_name} literal")
-            previous_was_digit = False
-            index += 1
-            continue
-
-        break
-
-    return index, saw_digit
-
-
-def _is_float_like_number(value: str) -> bool:
-    return "." in value
-
-
-def _read_string(source: str, start: int) -> str:
-    quote = source[start]
-    index = start + 1
-    escaped = False
-    while index < len(source):
-        char = source[index]
-        if escaped:
-            escaped = False
-        elif char == "\\":
-            escaped = True
-        elif char == quote:
-            return source[start:index + 1]
-        index += 1
-    raise GMLTranspileError("Unterminated string literal")
-
-
-def _normalize_local_names(local_names: Iterable[str] | None) -> frozenset[str]:
-    return frozenset(local_names or [])
-
-
-def _tokens_to_source(tokens: Iterable[_Token]) -> str:
-    return " ".join(token.value for token in tokens if token.kind not in ("EOF", "NEWLINE"))
-
-
-def _static_scope_id(
-    prefix: str,
-    name: str | None,
-    position: int,
-    body_tokens: Iterable[_Token],
-) -> str:
-    body_source = _tokens_to_source(body_tokens)
-    digest = hashlib.sha1(body_source.encode("utf-8")).hexdigest()[:12]
-    return f"{prefix}:{name or '<anonymous>'}:{position}:{digest}"
-
-
-def _collect_static_declarations(tokens: Iterable[_Token]) -> tuple[_StaticDeclaration, ...]:
-    token_list = list(tokens)
-    declarations: list[_StaticDeclaration] = []
-    index = 0
-    while index < len(token_list):
-        token = token_list[index]
-        if token.kind == "IDENT" and token.value == "function":
-            index = _skip_function_literal_tokens(token_list, index)
-            continue
-        if token.kind == "IDENT" and token.value == "static":
-            statement_tokens, index = _read_static_declaration_tokens(token_list, index + 1)
-            declarations.extend(_parse_static_declarations(_tokens_to_source(statement_tokens)))
-            continue
-        index += 1
-    return tuple(declarations)
-
-
-def _skip_function_literal_tokens(tokens: list[_Token], index: int) -> int:
-    body_start = index + 1
-    while body_start < len(tokens) and tokens[body_start].value != "{":
-        body_start += 1
-    if body_start >= len(tokens):
-        return index + 1
-
-    depth = 1
-    current = body_start + 1
-    while current < len(tokens) and depth > 0:
-        value = tokens[current].value
-        if value == "{":
-            depth += 1
-        elif value == "}":
-            depth -= 1
-        current += 1
-    return current
-
-
-def _read_static_declaration_tokens(
-    tokens: list[_Token],
-    index: int,
-) -> tuple[list[_Token], int]:
-    statement_tokens: list[_Token] = []
-    depth = 0
-    current = index
-    while current < len(tokens):
-        token = tokens[current]
-        if depth == 0 and token.value == ";":
-            return statement_tokens, current + 1
-        if depth == 0 and token.kind == "NEWLINE":
-            return statement_tokens, current + 1
-        if token.value in "([{":
-            depth += 1
-        elif token.value in ")]}" and depth > 0:
-            depth -= 1
-        statement_tokens.append(token)
-        current += 1
-    return statement_tokens, current
-
-
-def _parse_static_declarations(source: str) -> list[_StaticDeclaration]:
-    declarations: list[_StaticDeclaration] = []
-    for declaration in _split_top_level(source, ","):
-        declaration = declaration.strip()
-        if not declaration:
-            continue
-        assignment = _split_assignment(declaration)
-        if assignment is None:
-            name = declaration
-            value_source = "undefined"
-        else:
-            name, operator, value_source = assignment
-            if operator not in ("=", ":="):
-                raise GMLTranspileError("Static declarations only support simple assignments")
-            name = name.strip()
-            value_source = value_source.strip()
-        _validate_gml_identifier(name)
-        declarations.append(_StaticDeclaration(name, value_source))
-    return declarations
-
-
-def _emit_static_initialization_lines(
-    static_scope_name: str | None,
-    static_scope_id: str | None,
-    declarations: Iterable[_StaticDeclaration],
-    local_names: Iterable[str],
-    scope_context: _ScopeContext,
-    enum_values: MutableMapping[str, dict[str, int]],
-    enum_names: Iterable[str],
-    macro_values: Mapping[str, str],
-) -> list[str]:
-    if static_scope_name is None or static_scope_id is None:
-        return []
-
-    initializer_names = set(local_names)
-    initializers: list[str] = []
-    for declaration in declarations:
-        value = transpile_gml_expression(
-            declaration.value_source,
-            initializer_names,
-            enum_values,
-            enum_names,
-            scope_context=scope_context,
-            macro_values=macro_values,
-        )
-        initializers.append(f"[{json.dumps(declaration.name)}, func(): return {value}]")
-    return [
-        f"var {static_scope_name} = GMRuntime.gml_static_scope({json.dumps(static_scope_id)})",
-        f"GMRuntime.gml_static_initialize({static_scope_name}, [{', '.join(initializers)}])",
-    ]
-
-
-def _emit_constructor_inheritance_line(
-    parent_constructor: _Expression,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext,
-    constructor_scope_context: _ScopeContext,
-) -> str:
-    parent_expr = parent_constructor
-    args: tuple[_Expression, ...] = ()
-    if isinstance(parent_constructor, _Call):
-        parent_expr = parent_constructor.callee
-        args = parent_constructor.args
-
-    constructor = _emit_expression(
-        parent_expr,
-        local_names,
-        scope_context=constructor_scope_context,
-    )[0]
-    emitted_args = ", ".join(
-        _emit_expression(arg, local_names, scope_context=scope_context)[0]
-        for arg in args
-    )
-    return f"GMRuntime.gml_constructor_inherit(_gml_constructor_self, {constructor}, [{emitted_args}])"
-
-
-def _evaluate_enum_value_tokens(
-    tokens: Iterable[_Token],
-    enum_values: Mapping[str, Mapping[str, int]],
-    current_enum_values: Mapping[str, int],
-    macro_values: Mapping[str, str] | None = None,
-) -> int:
-    source = _tokens_to_source(tokens)
-    value = _evaluate_enum_expression(
-        _parse_gml_expression(source, macro_values=macro_values),
-        enum_values,
-        current_enum_values,
-    )
-    if value is None:
-        raise GMLTranspileError("Enum values must be integer compile-time constants")
-    return value
-
-
-def _evaluate_enum_expression(
-    expr: _Expression,
-    enum_values: Mapping[str, Mapping[str, int]],
-    current_enum_values: Mapping[str, int],
-) -> int | None:
-    if isinstance(expr, _NumberLiteral):
-        return _parse_enum_int_literal(expr)
-    if isinstance(expr, _Grouped):
-        return _evaluate_enum_expression(expr.expr, enum_values, current_enum_values)
-    if isinstance(expr, _Name):
-        if expr.value == "true":
-            return 1
-        if expr.value == "false":
-            return 0
-        return current_enum_values.get(expr.value)
-    if isinstance(expr, _Member) and isinstance(expr.target, _Name):
-        enum_members = enum_values.get(expr.target.value)
-        if enum_members is None:
-            return None
-        return enum_members.get(expr.member)
-    if isinstance(expr, _Unary):
-        operand = _evaluate_enum_expression(expr.operand, enum_values, current_enum_values)
-        if operand is None:
-            return None
-        if expr.operator == "+":
-            return operand
-        if expr.operator == "-":
-            return -operand
-        if expr.operator == "~":
-            return ~operand
-        return None
-    if isinstance(expr, _Binary):
-        left = _evaluate_enum_expression(expr.left, enum_values, current_enum_values)
-        right = _evaluate_enum_expression(expr.right, enum_values, current_enum_values)
-        if left is None or right is None:
-            return None
-        return _evaluate_enum_binary(expr.operator, left, right)
-    if isinstance(expr, _Call) and isinstance(expr.callee, _Name):
-        args: list[int] = []
-        for arg in expr.args:
-            arg_value = _evaluate_enum_expression(arg, enum_values, current_enum_values)
-            if arg_value is None:
-                return None
-            args.append(arg_value)
-        return _evaluate_enum_call(expr.callee.value, args)
-    return None
-
-
-def _parse_enum_int_literal(expr: _NumberLiteral) -> int | None:
-    if expr.is_float_like:
-        return None
-    try:
-        return int(expr.value, 0)
-    except ValueError:
-        return None
-
-
-def _evaluate_enum_binary(operator: str, left: int, right: int) -> int | None:
-    if operator == "+":
-        return left + right
-    if operator == "-":
-        return left - right
-    if operator == "*":
-        return left * right
-    if operator in ("div", "/"):
-        if right == 0:
-            return None
-        return int(left / right)
-    if operator in ("mod", "%"):
-        if right == 0:
-            return None
-        return left % right
-    if operator == "<<":
-        return left << right
-    if operator == ">>":
-        return left >> right
-    if operator == "&":
-        return left & right
-    if operator == "|":
-        return left | right
-    if operator == "^":
-        return left ^ right
-    return None
-
-
-def _evaluate_enum_call(function_name: str, args: Iterable[int]) -> int | None:
-    arg_values = list(args)
-    if function_name in ("int64", "real") and len(arg_values) == 1:
-        return int(arg_values[0])
-    if function_name == "bool" and len(arg_values) == 1:
-        return 1 if arg_values[0] else 0
-    if function_name == "abs" and len(arg_values) == 1:
-        return abs(int(arg_values[0]))
-    return None
-
-
-def _reject_enum_assignment_target(
-    target_expr: _Expression,
-    enum_names: Iterable[str] | None,
-) -> None:
-    enum_name_set = frozenset(enum_names or [])
-    if not enum_name_set:
-        return
-
-    unwrapped_target = _unwrap_grouped_expression(target_expr)
-    if _is_enum_reference(unwrapped_target, enum_name_set):
-        raise GMLTranspileError("Cannot assign to enum")
-    if _target_chain_starts_with_enum(unwrapped_target, enum_name_set):
-        raise GMLTranspileError("Cannot assign to enum member")
-
-
-def _reject_readonly_builtin_assignment_target(
-    target_expr: _Expression,
-    local_names: Iterable[str],
-) -> None:
-    local_name_set = _normalize_local_names(local_names)
-    unwrapped_target = _unwrap_grouped_expression(target_expr)
-    if isinstance(unwrapped_target, _Index):
-        unwrapped_target = _unwrap_grouped_expression(unwrapped_target.target)
-    if isinstance(unwrapped_target, _Name):
-        name = unwrapped_target.value
-        if name not in local_name_set and name in _READ_ONLY_BUILTIN_VARIABLES:
-            raise GMLTranspileError(f"Cannot assign to read-only built-in variable {name}")
-
-
-def _reject_constant_assignment_target_name(
-    target_source: str,
-    macro_names: Iterable[str],
-) -> None:
-    target_name = _raw_identifier_target_name(target_source)
-    if target_name is None:
-        return
-    if target_name in _GML_BUILTIN_CONSTANT_IDENTIFIERS:
-        raise GMLTranspileError(f"Cannot assign to built-in constant {target_name}")
-    if target_name in macro_names:
-        raise GMLTranspileError(f"Cannot assign to macro constant {target_name}")
-
-
-def _reject_constant_declaration_name(
-    name: str,
-    macro_names: Iterable[str],
-) -> None:
-    if name in _GML_BUILTIN_CONSTANT_IDENTIFIERS:
-        raise GMLTranspileError(f"Cannot redeclare built-in constant {name}")
-    if name in macro_names:
-        raise GMLTranspileError(f"Cannot redeclare macro constant {name}")
-
-
-def _raw_identifier_target_name(target_source: str) -> str | None:
-    tokens = _expression_tokens(target_source)
-    if len(tokens) == 2 and tokens[0].kind == "IDENT" and tokens[1].kind == "EOF":
-        return tokens[0].value
-    return None
-
-
-def _reject_enum_mutation_expression(
-    expr: _Expression,
-    enum_names: Iterable[str] | None,
-) -> None:
-    enum_name_set = frozenset(enum_names or [])
-    if not enum_name_set:
-        return
-
-    if isinstance(expr, _Call):
-        if (
-            isinstance(expr.callee, _Name)
-            and expr.callee.value in ("struct_set", "struct_remove")
-            and expr.args
-            and _is_enum_reference(expr.args[0], enum_name_set)
-        ):
-            raise GMLTranspileError("Cannot mutate enum member")
-        _reject_enum_mutation_expression(expr.callee, enum_name_set)
-        for arg in expr.args:
-            _reject_enum_mutation_expression(arg, enum_name_set)
-        return
-    if isinstance(expr, _Grouped):
-        _reject_enum_mutation_expression(expr.expr, enum_name_set)
-        return
-    if isinstance(expr, _Unary):
-        _reject_enum_mutation_expression(expr.operand, enum_name_set)
-        return
-    if isinstance(expr, _Binary):
-        _reject_enum_mutation_expression(expr.left, enum_name_set)
-        _reject_enum_mutation_expression(expr.right, enum_name_set)
-        return
-    if isinstance(expr, _Ternary):
-        _reject_enum_mutation_expression(expr.condition, enum_name_set)
-        _reject_enum_mutation_expression(expr.true_expr, enum_name_set)
-        _reject_enum_mutation_expression(expr.false_expr, enum_name_set)
-        return
-    if isinstance(expr, _ArrayLiteral):
-        for element in expr.elements:
-            _reject_enum_mutation_expression(element, enum_name_set)
-        return
-    if isinstance(expr, _StructLiteral):
-        for _field_name, field_value in expr.fields:
-            _reject_enum_mutation_expression(field_value, enum_name_set)
-        return
-    if isinstance(expr, _Index):
-        _reject_enum_mutation_expression(expr.target, enum_name_set)
-        _reject_enum_mutation_expression(expr.index, enum_name_set)
-        return
-    if isinstance(expr, _StructAccess):
-        _reject_enum_mutation_expression(expr.target, enum_name_set)
-        _reject_enum_mutation_expression(expr.key, enum_name_set)
-        return
-    if isinstance(expr, _DSMapAccess):
-        _reject_enum_mutation_expression(expr.target, enum_name_set)
-        _reject_enum_mutation_expression(expr.key, enum_name_set)
-        return
-    if isinstance(expr, _Member):
-        _reject_enum_mutation_expression(expr.target, enum_name_set)
-
-
-def _unwrap_grouped_expression(expr: _Expression) -> _Expression:
-    while isinstance(expr, _Grouped):
-        expr = expr.expr
-    return expr
-
-
-def _is_enum_reference(expr: _Expression, enum_names: Iterable[str]) -> bool:
-    unwrapped_expr = _unwrap_grouped_expression(expr)
-    return isinstance(unwrapped_expr, _Name) and unwrapped_expr.value in enum_names
-
-
-def _target_chain_starts_with_enum(expr: _Expression, enum_names: Iterable[str]) -> bool:
-    unwrapped_expr = _unwrap_grouped_expression(expr)
-    if isinstance(unwrapped_expr, _Member | _StructAccess | _DSMapAccess | _Index):
-        return _is_enum_reference(unwrapped_expr.target, enum_names) or _target_chain_starts_with_enum(
-            unwrapped_expr.target,
-            enum_names,
-        )
-    return False
-
-
-def _split_top_level_tokens(tokens: Iterable[_Token], separator: str) -> list[list[_Token]]:
-    parts: list[list[_Token]] = [[]]
-    depth = 0
-    for token in tokens:
-        if token.value in "([{":
-            depth += 1
-        elif token.value in ")]}" and depth > 0:
-            depth -= 1
-
-        if depth == 0 and token.value == separator:
-            parts.append([])
-            continue
-        parts[-1].append(token)
-    return parts
-
-
-def _indent_lines(lines: Iterable[str]) -> list[str]:
-    return [f"\t{line}" if line else "" for line in lines]
-
-
-def _insert_lines_before_continue(lines: Iterable[str], inserted_lines: Iterable[str]) -> list[str]:
-    inserted = list(inserted_lines)
-    result: list[str] = []
-    for line in lines:
-        stripped = line.lstrip("\t")
-        if stripped == "continue":
-            indentation = line[: len(line) - len(stripped)]
-            result.extend(f"{indentation}{inserted_line}" for inserted_line in inserted)
-        result.append(line)
-    return result
-
-
-def _insert_until_check_before_continue(lines: Iterable[str], condition: str) -> list[str]:
-    result: list[str] = []
-    for line in lines:
-        stripped = line.lstrip("\t")
-        if stripped == "continue":
-            indentation = line[: len(line) - len(stripped)]
-            result.append(f"{indentation}if {condition}:")
-            result.append(f"{indentation}\tbreak")
-        result.append(line)
-    return result
-
-
-def _next_generated_name_from_counter(generated_counter: list[int], prefix: str) -> str:
-    index = generated_counter[0]
-    generated_counter[0] += 1
-    return f"{prefix}_{index}"
-
-
-def _expression_needs_assignment_cache(expr: _Expression) -> bool:
-    if isinstance(expr, _Call | _NewCall | _FunctionLiteral):
-        return True
-    if isinstance(expr, _Grouped):
-        return _expression_needs_assignment_cache(expr.expr)
-    if isinstance(expr, _Unary):
-        return _expression_needs_assignment_cache(expr.operand)
-    if isinstance(expr, _Binary):
-        return (
-            _expression_needs_assignment_cache(expr.left)
-            or _expression_needs_assignment_cache(expr.right)
-        )
-    if isinstance(expr, _Ternary):
-        return (
-            _expression_needs_assignment_cache(expr.condition)
-            or _expression_needs_assignment_cache(expr.true_expr)
-            or _expression_needs_assignment_cache(expr.false_expr)
-        )
-    if isinstance(expr, _ArrayLiteral):
-        return any(_expression_needs_assignment_cache(element) for element in expr.elements)
-    if isinstance(expr, _StructLiteral):
-        return any(
-            _expression_needs_assignment_cache(field_value)
-            for _field_name, field_value in expr.fields
-        )
-    if isinstance(expr, _Index):
-        return (
-            _expression_needs_assignment_cache(expr.target)
-            or _expression_needs_assignment_cache(expr.index)
-        )
-    if isinstance(expr, _StructAccess):
-        return (
-            _expression_needs_assignment_cache(expr.target)
-            or _expression_needs_assignment_cache(expr.key)
-        )
-    if isinstance(expr, _DSMapAccess):
-        return (
-            _expression_needs_assignment_cache(expr.target)
-            or _expression_needs_assignment_cache(expr.key)
-        )
-    if isinstance(expr, _Member):
-        return _expression_needs_assignment_cache(expr.target)
-    return False
-
-
-def _cache_assignment_part(
-    prelude_lines: list[str],
-    expr: _Expression,
-    emitted: str,
-    generated_counter: list[int],
-    prefix: str,
-) -> str:
-    if not _expression_needs_assignment_cache(expr):
-        return emitted
-    temp_name = _next_generated_name_from_counter(generated_counter, prefix)
-    prelude_lines.append(f"var {temp_name} = {emitted}")
-    return temp_name
-
-
-def _normalize_scope_context(scope_context: _ScopeContext | None) -> _ScopeContext:
-    return scope_context if scope_context is not None else _DEFAULT_SCOPE_CONTEXT
-
-
-def _scope_context_with_global_names(
-    scope_context: _ScopeContext,
-    global_names: Iterable[str] | None,
-    top_level_global_scope: bool | None = None,
-    asset_names: Iterable[str] | None = None,
-    static_prefix: str | None = None,
-) -> _ScopeContext:
-    names = set(scope_context.global_names)
-    names.update(global_names or [])
-    assets = set(scope_context.asset_names)
-    assets.update(asset_names or [])
-    return _ScopeContext(
-        self_expression=scope_context.self_expression,
-        other_expression=scope_context.other_expression,
-        instance_target=scope_context.instance_target,
-        global_scope=scope_context.global_scope if top_level_global_scope is None else top_level_global_scope,
-        global_names=frozenset(names),
-        asset_names=frozenset(assets),
-        static_scope=scope_context.static_scope,
-        static_names=scope_context.static_names,
-        static_prefix=scope_context.static_prefix if static_prefix is None else static_prefix,
-    )
-
-
-def _macro_configuration_matches(configuration: str, active_configuration: str | None) -> bool:
-    if active_configuration is None:
-        return False
-    return configuration.casefold() == active_configuration.casefold()
-
-
-def _emit_name(
-    value: str,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext,
-) -> tuple[str, int]:
-    is_local = value in local_names
-    if not is_local and value == "self":
-        return scope_context.self_expression, _PRIMARY_PRECEDENCE
-    if not is_local and value == "other":
-        return scope_context.other_expression, _PRIMARY_PRECEDENCE
-    if not is_local and value in _GML_LITERAL_IDENTIFIERS:
-        return value, _PRIMARY_PRECEDENCE
-    if (
-        not is_local
-        and scope_context.static_scope is not None
-        and value in scope_context.static_names
-    ):
-        return (
-            f"GMRuntime.gml_struct_get({scope_context.static_scope}, {json.dumps(value)})",
-            _POSTFIX_PRECEDENCE,
-        )
-    if not is_local:
-        if value in _BUILTIN_ARRAY_VARIABLES:
-            return f"GMRuntime.gml_builtin_array({json.dumps(value)})", _POSTFIX_PRECEDENCE
-        if value in _BUILTIN_GLOBAL_VARIABLES:
-            return f"GMRuntime.gml_builtin_global({json.dumps(value)})", _POSTFIX_PRECEDENCE
-        if _name_resolves_to_global(value, local_names, scope_context):
-            return (
-                "GMRuntime.gml_struct_get("
-                f"GMRuntime.gml_global_scope(), {json.dumps(value)})"
-            ), _POSTFIX_PRECEDENCE
-        if scope_context.instance_target is not None and _is_plain_identifier(value):
-            return (
-                "GMRuntime.gml_variable_instance_get("
-                f"{scope_context.instance_target}, {json.dumps(value)})"
-            ), _POSTFIX_PRECEDENCE
-        value = _INSTANCE_NAME_REPLACEMENTS.get(value, value)
-    value = _sanitize_gdscript_identifier(value)
-    return value, _PRIMARY_PRECEDENCE
-
-
-def _name_resolves_to_global(
-    name: str,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext,
-) -> bool:
-    if name in local_names or not _is_plain_identifier(name):
-        return False
-    if name in _GML_LITERAL_IDENTIFIERS:
-        return False
-    if name in _BUILTIN_ARRAY_VARIABLES or name in _BUILTIN_GLOBAL_VARIABLES:
-        return False
-    if name in scope_context.global_names:
-        return True
-    return (
-        scope_context.global_scope
-        and name not in _BUILTIN_INSTANCE_VARIABLES
-        and name not in _GML_BUILTIN_CONSTANT_IDENTIFIERS
-    )
-
-
-def _emit_expression(
-    expr: _Expression,
-    local_names: Iterable[str] | None = None,
-    bind_function_literals: bool = True,
-    scope_context: _ScopeContext | None = None,
-) -> tuple[str, int]:
-    local_names = _normalize_local_names(local_names)
-    scope_context = _normalize_scope_context(scope_context)
-    if isinstance(expr, _Literal | _StringLiteral | _NumberLiteral):
-        return expr.value, _PRIMARY_PRECEDENCE
-    if isinstance(expr, _NameOf):
-        return json.dumps(expr.value), _PRIMARY_PRECEDENCE
-    if isinstance(expr, _Name):
-        return _emit_name(expr.value, local_names, scope_context)
-    if isinstance(expr, _Grouped):
-        return (
-            f"({_emit_expression(expr.expr, local_names, scope_context=scope_context)[0]})",
-            _PRIMARY_PRECEDENCE,
-        )
-    if isinstance(expr, _Unary):
-        if expr.operator == "!":
-            return (
-                f"not {_emit_truthy_expression(expr.operand, local_names, scope_context=scope_context)}",
-                _UNARY_PRECEDENCE,
-            )
-        if expr.operator == "not":
-            return (
-                f"not {_emit_truthy_expression(expr.operand, local_names, scope_context=scope_context)}",
-                _UNARY_PRECEDENCE,
-            )
-        if expr.operator == "~":
-            operand = _emit_expression(expr.operand, local_names, scope_context=scope_context)[0]
-            return f"GMRuntime.gml_bit_not({operand})", _POSTFIX_PRECEDENCE
-        operand = _emit_child(
-            expr.operand,
-            _UNARY_PRECEDENCE,
-            local_names=local_names,
-            scope_context=scope_context,
-        )
-        return f"{expr.operator}{operand}", _UNARY_PRECEDENCE
-    if isinstance(expr, _Binary):
-        return _emit_binary(expr, local_names, scope_context=scope_context)
-    if isinstance(expr, _Ternary):
-        condition = _emit_truthy_expression(
-            expr.condition,
-            local_names,
-            scope_context=scope_context,
-        )
-        true_expr = _emit_child(
-            expr.true_expr,
-            _TERNARY_PRECEDENCE,
-            local_names=local_names,
-            scope_context=scope_context,
-        )
-        false_expr = _emit_child(
-            expr.false_expr,
-            _TERNARY_PRECEDENCE,
-            local_names=local_names,
-            scope_context=scope_context,
-        )
-        return f"{true_expr} if {condition} else {false_expr}", _TERNARY_PRECEDENCE
-    if isinstance(expr, _Call):
-        builtin_call = _emit_builtin_call(expr, local_names, scope_context=scope_context)
-        if builtin_call is not None:
-            return builtin_call, _POSTFIX_PRECEDENCE
-        callee = _emit_child(
-            expr.callee,
-            _POSTFIX_PRECEDENCE,
-            local_names=local_names,
-            scope_context=scope_context,
-        )
-        args = ", ".join(
-            _emit_expression(arg, local_names, scope_context=scope_context)[0]
-            for arg in expr.args
-        )
-        return f"{callee}({args})", _POSTFIX_PRECEDENCE
-    if isinstance(expr, _ArrayLiteral):
-        elements = ", ".join(
-            _emit_expression(element, local_names, scope_context=scope_context)[0]
-            for element in expr.elements
-        )
-        return f"[{elements}]", _PRIMARY_PRECEDENCE
-    if isinstance(expr, _FunctionLiteral):
-        function_literal = _emit_function_literal(expr, local_names, scope_context=scope_context)
-        if expr.static_scope_id is not None:
-            function_literal = (
-                "GMRuntime.gml_static_bind("
-                f"{function_literal}, {json.dumps(expr.static_scope_id)}, {json.dumps(expr.name or '')})"
-            )
-        if bind_function_literals:
-            if expr.is_constructor:
-                return (
-                    f"GMRuntime.gml_constructor({scope_context.self_expression}, {function_literal})",
-                    _POSTFIX_PRECEDENCE,
-                )
-            return (
-                f"GMRuntime.gml_method({scope_context.self_expression}, {function_literal})",
-                _POSTFIX_PRECEDENCE,
-            )
-        return function_literal, _PRIMARY_PRECEDENCE
-    if isinstance(expr, _NewCall):
-        constructor = _emit_expression(
-            expr.constructor,
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        args = ", ".join(
-            _emit_expression(arg, local_names, scope_context=scope_context)[0]
-            for arg in expr.args
-        )
-        return f"GMRuntime.gml_new({constructor}, [{args}])", _POSTFIX_PRECEDENCE
-    if isinstance(expr, _StructLiteral):
-        fields = ", ".join(
-            _emit_struct_field(
-                field_name,
-                field_value,
-                local_names,
-                scope_context=scope_context,
-            )
-            for field_name, field_value in expr.fields
-        )
-        return f"GMRuntime.gml_struct({{{fields}}})", _POSTFIX_PRECEDENCE
-    if isinstance(expr, _Index):
-        target = _emit_expression(expr.target, local_names, scope_context=scope_context)[0]
-        index = _emit_expression(expr.index, local_names, scope_context=scope_context)[0]
-        return f"GMRuntime.gml_array_get({target}, {index})", _POSTFIX_PRECEDENCE
-    if isinstance(expr, _StructAccess):
-        target = _emit_expression(expr.target, local_names, scope_context=scope_context)[0]
-        key = _emit_expression(expr.key, local_names, scope_context=scope_context)[0]
-        return f"GMRuntime.gml_struct_get({target}, {key})", _POSTFIX_PRECEDENCE
-    if isinstance(expr, _DSMapAccess):
-        target = _emit_expression(expr.target, local_names, scope_context=scope_context)[0]
-        key = _emit_expression(expr.key, local_names, scope_context=scope_context)[0]
-        return f"GMRuntime.gml_ds_map_find_value({target}, {key})", _POSTFIX_PRECEDENCE
-    if _uses_direct_member_access(expr, scope_context=scope_context):
-        target = _emit_child(
-            expr.target,
-            _POSTFIX_PRECEDENCE,
-            local_names=local_names,
-            scope_context=scope_context,
-        )
-        return f"{target}.{_sanitize_gdscript_identifier(expr.member)}", _POSTFIX_PRECEDENCE
-    target = _emit_child(
-        expr.target,
-        _POSTFIX_PRECEDENCE,
-        local_names=local_names,
-        scope_context=scope_context,
-    )
-    return f"GMRuntime.gml_struct_get({target}, {json.dumps(expr.member)})", _POSTFIX_PRECEDENCE
-
-
-def _uses_direct_member_access(
-    expr: _Member,
-    scope_context: _ScopeContext | None = None,
-) -> bool:
-    scope_context = _normalize_scope_context(scope_context)
-    if not isinstance(expr.target, _Name):
-        return False
-    if expr.target.value == "self" and scope_context.self_expression != "self":
-        return False
-    if expr.target.value == "other" and scope_context.other_expression != "other":
-        return False
-    return expr.target.value in _DIRECT_MEMBER_TARGETS
-
-
-def _emit_function_literal(
-    expr: _FunctionLiteral,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> str:
-    name = f" {_sanitize_gdscript_identifier(expr.name)}" if expr.name is not None else ""
-    parameter_names = [parameter.name for parameter in expr.parameters]
-    emitted_parameters = [
-        f"{_sanitize_gdscript_identifier(parameter.name)} = null"
-        for parameter in expr.parameters
-    ]
-    if expr.is_constructor:
-        emitted_parameters.insert(0, "_gml_constructor_self = null")
-    parameters = ", ".join(emitted_parameters)
-    default_lines = _emit_function_parameter_default_lines(
-        expr.parameters,
-        parameter_names,
-        scope_context=scope_context,
-    )
-    body = "; ".join([*default_lines, *expr.body_lines])
-    return f"func{name}({parameters}): {body}"
-
-
-def _emit_struct_field(
-    field_name: str,
-    field_value: _Expression,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> str:
-    bind_function_literal = not isinstance(_unwrap_grouped_expression(field_value), _FunctionLiteral)
-    field_text = _emit_expression(
-        field_value,
-        local_names,
-        bind_function_literal,
-        scope_context=scope_context,
-    )[0]
-    return f"{json.dumps(field_name)}: {field_text}"
-
-
-def _emit_function_parameter_default_lines(
-    parameters: Iterable[_FunctionParameter],
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> list[str]:
-    default_lines: list[str] = []
-    for parameter in parameters:
-        parameter_name = _sanitize_gdscript_identifier(parameter.name)
-        if parameter.default is None:
-            default_lines.append(f"if {parameter_name} == null: {parameter_name} = GMRuntime.gml_undefined()")
-            continue
-        default_value = _emit_expression(
-            parameter.default,
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        default_lines.append(
-            f"if {parameter_name} == null or GMRuntime.is_undefined({parameter_name}): "
-            f"{parameter_name} = {default_value}",
-        )
-    return default_lines
-
-
-def _emit_builtin_call(
-    expr: _Call,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> str | None:
-    scope_context = _normalize_scope_context(scope_context)
-    if isinstance(expr.callee, _Name) and expr.callee.value == "keyboard_check" and len(expr.args) == 1:
-        key = expr.args[0]
-        if isinstance(key, _Name) and key.value in _VIRTUAL_KEY_ACTIONS:
-            return f'Input.is_action_pressed("{_VIRTUAL_KEY_ACTIONS[key.value]}")'
-        if isinstance(key, _Name) and key.value in _VIRTUAL_KEY_CONSTANTS:
-            return f"Input.is_key_pressed({_VIRTUAL_KEY_CONSTANTS[key.value]})"
-    if isinstance(expr.callee, _Name) and expr.callee.value == "method" and len(expr.args) == 2:
-        scope = _emit_expression(expr.args[0], local_names, scope_context=scope_context)[0]
-        if scope == _NAME_REPLACEMENTS["undefined"]:
-            scope = scope_context.self_expression
-        function_value = _emit_expression(
-            expr.args[1],
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        return f"GMRuntime.gml_method({scope}, {function_value})"
-    if isinstance(expr.callee, _Name) and expr.callee.value == "with_targets" and len(expr.args) == 1:
-        target = _emit_instance_keyword_argument(
-            expr.args[0],
-            local_names,
-            scope_context=scope_context,
-        )
-        return f"GMRuntime.gml_with_targets({target})"
-    if isinstance(expr.callee, _Name) and expr.callee.value in _STRUCT_RUNTIME_FUNCTIONS:
-        args = ", ".join(
-            _emit_expression(arg, local_names, scope_context=scope_context)[0]
-            for arg in expr.args
-        )
-        return f"GMRuntime.{_STRUCT_RUNTIME_FUNCTIONS[expr.callee.value]}({args})"
-    if (
-        isinstance(expr.callee, _Name)
-        and expr.callee.value in _VARIABLE_RUNTIME_FUNCTIONS
-        and expr.callee.value.startswith("variable_instance_")
-        and expr.args
-    ):
-        first_arg = _emit_instance_keyword_argument(
-            expr.args[0],
-            local_names,
-            scope_context=scope_context,
-        )
-        remaining_args = [
-            _emit_expression(arg, local_names, scope_context=scope_context)[0]
-            for arg in expr.args[1:]
-        ]
-        args = ", ".join([first_arg, *remaining_args])
-        return f"GMRuntime.{_VARIABLE_RUNTIME_FUNCTIONS[expr.callee.value]}({args})"
-    if isinstance(expr.callee, _Name) and expr.callee.value in _VARIABLE_RUNTIME_FUNCTIONS:
-        args = ", ".join(
-            _emit_expression(arg, local_names, scope_context=scope_context)[0]
-            for arg in expr.args
-        )
-        return f"GMRuntime.{_VARIABLE_RUNTIME_FUNCTIONS[expr.callee.value]}({args})"
-    if isinstance(expr.callee, _Name) and expr.callee.value in _DS_MAP_RUNTIME_FUNCTIONS:
-        args = ", ".join(
-            _emit_expression(arg, local_names, scope_context=scope_context)[0]
-            for arg in expr.args
-        )
-        return f"GMRuntime.{_DS_MAP_RUNTIME_FUNCTIONS[expr.callee.value]}({args})"
-    if isinstance(expr.callee, _Name) and expr.callee.value in _ARRAY_RUNTIME_FUNCTIONS:
-        args = ", ".join(
-            _emit_expression(arg, local_names, scope_context=scope_context)[0]
-            for arg in expr.args
-        )
-        return f"GMRuntime.{_ARRAY_RUNTIME_FUNCTIONS[expr.callee.value]}({args})"
-    if (
-        isinstance(expr.callee, _Name)
-        and expr.callee.value in _RUNTIME_FUNCTIONS
-        and len(expr.args) == 1
-    ):
-        arg = _emit_expression(expr.args[0], local_names, scope_context=scope_context)[0]
-        return f"GMRuntime.{_RUNTIME_FUNCTIONS[expr.callee.value]}({arg})"
-    return None
-
-
-def _emit_instance_keyword_argument(
-    expr: _Expression,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> str:
-    scope_context = _normalize_scope_context(scope_context)
-    legacy_keyword = _legacy_instance_keyword_value(expr)
-    if legacy_keyword == -1:
-        return scope_context.self_expression
-    if legacy_keyword == -2:
-        return scope_context.other_expression
-    if legacy_keyword == -3:
-        return "GMRuntime.gml_instance_all()"
-    if legacy_keyword == -4:
-        return "GMRuntime.gml_instance_noone()"
-    return _emit_expression(expr, local_names, scope_context=scope_context)[0]
-
-
-def _legacy_instance_keyword_value(expr: _Expression) -> int | None:
-    unwrapped_expr = _unwrap_grouped_expression(expr)
-    if not isinstance(unwrapped_expr, _Unary) or unwrapped_expr.operator != "-":
-        return None
-    operand = _unwrap_grouped_expression(unwrapped_expr.operand)
-    if not isinstance(operand, _NumberLiteral) or operand.is_float_like:
-        return None
-    try:
-        value = int(operand.value, 0)
-    except ValueError:
-        return None
-    if value in (1, 2, 3, 4):
-        return -value
-    return None
-
-
-def _emit_binary(
-    expr: _Binary,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> tuple[str, int]:
-    scope_context = _normalize_scope_context(scope_context)
-    operator = _OPERATOR_REPLACEMENTS.get(expr.operator, expr.operator)
-
-    if expr.operator in ("&&", "and", "||", "or"):
-        operator = "and" if expr.operator in ("&&", "and") else "or"
-        left = _emit_truthy_expression(expr.left, local_names, scope_context=scope_context)
-        right = _emit_truthy_expression(expr.right, local_names, scope_context=scope_context)
-        return f"{left} {operator} {right}", _BINARY_PRECEDENCE[expr.operator]
-
-    if expr.operator == "^^":
-        left = _emit_truthy_expression(expr.left, local_names, scope_context=scope_context)
-        right = _emit_truthy_expression(expr.right, local_names, scope_context=scope_context)
-        return f"{left} != {right}", _BINARY_PRECEDENCE[expr.operator]
-
-    if expr.operator == "div":
-        left = _emit_expression(expr.left, local_names, scope_context=scope_context)[0]
-        right = _emit_expression(expr.right, local_names, scope_context=scope_context)[0]
-        return f"GMRuntime.gml_int_div({left}, {right})", _POSTFIX_PRECEDENCE
-
-    if expr.operator == "??":
-        left = _emit_expression(expr.left, local_names, scope_context=scope_context)[0]
-        right = _emit_child(
-            expr.right,
-            _TERNARY_PRECEDENCE,
-            local_names=local_names,
-            scope_context=scope_context,
-        )
-        return f"{left} if not GMRuntime.gml_is_nullish({left}) else {right}", _TERNARY_PRECEDENCE
-
-    if expr.operator in ("=", "==", "!=") and (
-        _contains_gml_undefined(expr.left)
-        or _contains_gml_undefined(expr.right)
-        or _contains_gml_nan(expr.left)
-        or _contains_gml_nan(expr.right)
-        or _contains_gml_pointer(expr.left)
-        or _contains_gml_pointer(expr.right)
-        or _contains_gml_handle(expr.left)
-        or _contains_gml_handle(expr.right)
-        or _may_need_gml_reference_equality(expr.left)
-        or _may_need_gml_reference_equality(expr.right)
-    ):
-        left = _emit_expression(expr.left, local_names, scope_context=scope_context)[0]
-        right = _emit_expression(expr.right, local_names, scope_context=scope_context)[0]
-        helper = "gml_ne" if expr.operator == "!=" else "gml_eq"
-        return f"GMRuntime.{helper}({left}, {right})", _POSTFIX_PRECEDENCE
-
-    if expr.operator == "/":
-        left = _emit_expression(expr.left, local_names, scope_context=scope_context)[0]
-        right = _emit_expression(expr.right, local_names, scope_context=scope_context)[0]
-        return f"GMRuntime.gml_div({left}, {right})", _POSTFIX_PRECEDENCE
-
-    if expr.operator in _ARITHMETIC_RUNTIME_FUNCTIONS:
-        left = _emit_expression(expr.left, local_names, scope_context=scope_context)[0]
-        right = _emit_expression(expr.right, local_names, scope_context=scope_context)[0]
-        return f"GMRuntime.{_ARITHMETIC_RUNTIME_FUNCTIONS[expr.operator]}({left}, {right})", _POSTFIX_PRECEDENCE
-
-    if expr.operator in _BITWISE_RUNTIME_FUNCTIONS:
-        left = _emit_expression(expr.left, local_names, scope_context=scope_context)[0]
-        right = _emit_expression(expr.right, local_names, scope_context=scope_context)[0]
-        return f"GMRuntime.{_BITWISE_RUNTIME_FUNCTIONS[expr.operator]}({left}, {right})", _POSTFIX_PRECEDENCE
-
-    precedence = _BINARY_PRECEDENCE[expr.operator]
-    left = _emit_child(
-        expr.left,
-        precedence,
-        local_names=local_names,
-        scope_context=scope_context,
-    )
-    right = _emit_child(
-        expr.right,
-        precedence,
-        is_right_child=True,
-        parent_operator=expr.operator,
-        local_names=local_names,
-        scope_context=scope_context,
-    )
-    return f"{left} {operator} {right}", precedence
-
-
-def _contains_gml_undefined(expr: _Expression) -> bool:
-    if isinstance(expr, _Name):
-        return expr.value == "GMRuntime.gml_undefined()"
-    if isinstance(expr, _Grouped):
-        return _contains_gml_undefined(expr.expr)
-    if isinstance(expr, _Unary):
-        return _contains_gml_undefined(expr.operand)
-    if isinstance(expr, _Binary):
-        return _contains_gml_undefined(expr.left) or _contains_gml_undefined(expr.right)
-    if isinstance(expr, _Ternary):
-        return (
-            _contains_gml_undefined(expr.condition)
-            or _contains_gml_undefined(expr.true_expr)
-            or _contains_gml_undefined(expr.false_expr)
-        )
-    if isinstance(expr, _Call):
-        return _contains_gml_undefined(expr.callee) or any(
-            _contains_gml_undefined(arg) for arg in expr.args
-        )
-    if isinstance(expr, _ArrayLiteral):
-        return any(_contains_gml_undefined(element) for element in expr.elements)
-    if isinstance(expr, _FunctionLiteral):
-        return False
-    if isinstance(expr, _NewCall):
-        return _contains_gml_undefined(expr.constructor) or any(
-            _contains_gml_undefined(arg) for arg in expr.args
-        )
-    if isinstance(expr, _StructLiteral):
-        return any(_contains_gml_undefined(field_value) for _field_name, field_value in expr.fields)
-    if isinstance(expr, _Index):
-        return _contains_gml_undefined(expr.target) or _contains_gml_undefined(expr.index)
-    if isinstance(expr, _StructAccess):
-        return _contains_gml_undefined(expr.target) or _contains_gml_undefined(expr.key)
-    if isinstance(expr, _DSMapAccess):
-        return _contains_gml_undefined(expr.target) or _contains_gml_undefined(expr.key)
-    if isinstance(expr, _Member):
-        return _contains_gml_undefined(expr.target)
-    return False
-
-
-def _contains_gml_nan(expr: _Expression) -> bool:
-    if isinstance(expr, _Name):
-        return expr.value == "NAN"
-    if isinstance(expr, _Grouped):
-        return _contains_gml_nan(expr.expr)
-    if isinstance(expr, _Unary):
-        return _contains_gml_nan(expr.operand)
-    if isinstance(expr, _Binary):
-        return _contains_gml_nan(expr.left) or _contains_gml_nan(expr.right)
-    if isinstance(expr, _Ternary):
-        return (
-            _contains_gml_nan(expr.condition)
-            or _contains_gml_nan(expr.true_expr)
-            or _contains_gml_nan(expr.false_expr)
-        )
-    if isinstance(expr, _Call):
-        return _contains_gml_nan(expr.callee) or any(_contains_gml_nan(arg) for arg in expr.args)
-    if isinstance(expr, _ArrayLiteral):
-        return any(_contains_gml_nan(element) for element in expr.elements)
-    if isinstance(expr, _FunctionLiteral):
-        return False
-    if isinstance(expr, _NewCall):
-        return _contains_gml_nan(expr.constructor) or any(
-            _contains_gml_nan(arg) for arg in expr.args
-        )
-    if isinstance(expr, _StructLiteral):
-        return any(_contains_gml_nan(field_value) for _field_name, field_value in expr.fields)
-    if isinstance(expr, _Index):
-        return _contains_gml_nan(expr.target) or _contains_gml_nan(expr.index)
-    if isinstance(expr, _StructAccess):
-        return _contains_gml_nan(expr.target) or _contains_gml_nan(expr.key)
-    if isinstance(expr, _DSMapAccess):
-        return _contains_gml_nan(expr.target) or _contains_gml_nan(expr.key)
-    if isinstance(expr, _Member):
-        return _contains_gml_nan(expr.target)
-    return False
-
-
-def _contains_gml_pointer(expr: _Expression) -> bool:
-    if isinstance(expr, _Name):
-        return expr.value in ("GMRuntime.gml_pointer_null()", "GMRuntime.gml_pointer_invalid()")
-    if isinstance(expr, _Grouped):
-        return _contains_gml_pointer(expr.expr)
-    if isinstance(expr, _Unary):
-        return _contains_gml_pointer(expr.operand)
-    if isinstance(expr, _Binary):
-        return _contains_gml_pointer(expr.left) or _contains_gml_pointer(expr.right)
-    if isinstance(expr, _Ternary):
-        return (
-            _contains_gml_pointer(expr.condition)
-            or _contains_gml_pointer(expr.true_expr)
-            or _contains_gml_pointer(expr.false_expr)
-        )
-    if isinstance(expr, _Call):
-        return (
-            isinstance(expr.callee, _Name)
-            and expr.callee.value == "ptr"
-        ) or any(_contains_gml_pointer(arg) for arg in expr.args)
-    if isinstance(expr, _ArrayLiteral):
-        return any(_contains_gml_pointer(element) for element in expr.elements)
-    if isinstance(expr, _FunctionLiteral):
-        return False
-    if isinstance(expr, _NewCall):
-        return _contains_gml_pointer(expr.constructor) or any(
-            _contains_gml_pointer(arg) for arg in expr.args
-        )
-    if isinstance(expr, _StructLiteral):
-        return any(_contains_gml_pointer(field_value) for _field_name, field_value in expr.fields)
-    if isinstance(expr, _Index):
-        return _contains_gml_pointer(expr.target) or _contains_gml_pointer(expr.index)
-    if isinstance(expr, _StructAccess):
-        return _contains_gml_pointer(expr.target) or _contains_gml_pointer(expr.key)
-    if isinstance(expr, _DSMapAccess):
-        return _contains_gml_pointer(expr.target) or _contains_gml_pointer(expr.key)
-    if isinstance(expr, _Member):
-        return _contains_gml_pointer(expr.target)
-    return False
-
-
-def _contains_gml_handle(expr: _Expression) -> bool:
-    if isinstance(expr, _Name):
-        return expr.value == "GMRuntime.gml_instance_noone()"
-    if isinstance(expr, _Grouped):
-        return _contains_gml_handle(expr.expr)
-    if isinstance(expr, _Unary):
-        return _contains_gml_handle(expr.operand)
-    if isinstance(expr, _Binary):
-        return _contains_gml_handle(expr.left) or _contains_gml_handle(expr.right)
-    if isinstance(expr, _Ternary):
-        return (
-            _contains_gml_handle(expr.condition)
-            or _contains_gml_handle(expr.true_expr)
-            or _contains_gml_handle(expr.false_expr)
-        )
-    if isinstance(expr, _Call):
-        return any(_contains_gml_handle(arg) for arg in expr.args)
-    if isinstance(expr, _ArrayLiteral):
-        return any(_contains_gml_handle(element) for element in expr.elements)
-    if isinstance(expr, _FunctionLiteral):
-        return False
-    if isinstance(expr, _NewCall):
-        return _contains_gml_handle(expr.constructor) or any(
-            _contains_gml_handle(arg) for arg in expr.args
-        )
-    if isinstance(expr, _StructLiteral):
-        return any(_contains_gml_handle(field_value) for _field_name, field_value in expr.fields)
-    if isinstance(expr, _Index):
-        return _contains_gml_handle(expr.target) or _contains_gml_handle(expr.index)
-    if isinstance(expr, _StructAccess):
-        return _contains_gml_handle(expr.target) or _contains_gml_handle(expr.key)
-    if isinstance(expr, _DSMapAccess):
-        return _contains_gml_handle(expr.target) or _contains_gml_handle(expr.key)
-    if isinstance(expr, _Member):
-        return _contains_gml_handle(expr.target)
-    return False
-
-
-def _may_need_gml_reference_equality(expr: _Expression) -> bool:
-    if isinstance(expr, _Name):
-        return expr.value not in {"true", "false", "INF", "NAN"}
-    if isinstance(expr, (_ArrayLiteral, _StructLiteral, _FunctionLiteral)):
-        return True
-    if isinstance(expr, (_Call, _NewCall, _Index, _StructAccess, _DSMapAccess, _Member)):
-        return True
-    if isinstance(expr, _Grouped):
-        return _may_need_gml_reference_equality(expr.expr)
-    if isinstance(expr, _Unary):
-        return _may_need_gml_reference_equality(expr.operand)
-    if isinstance(expr, _Binary):
-        return _may_need_gml_reference_equality(expr.left) or _may_need_gml_reference_equality(expr.right)
-    if isinstance(expr, _Ternary):
-        return (
-            _may_need_gml_reference_equality(expr.condition)
-            or _may_need_gml_reference_equality(expr.true_expr)
-            or _may_need_gml_reference_equality(expr.false_expr)
-        )
-    return False
-
-
-def _emit_truthy_expression(
-    expr: _Expression,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> str:
-    if _emits_boolean_result(expr):
-        return _emit_expression(expr, local_names, scope_context=scope_context)[0]
-    return _gml_bool_call(_emit_expression(expr, local_names, scope_context=scope_context)[0])
-
-
-def _emits_boolean_result(expr: _Expression) -> bool:
-    if isinstance(expr, _Name):
-        return expr.value in ("true", "false")
-    if isinstance(expr, _Grouped):
-        return _emits_boolean_result(expr.expr)
-    if isinstance(expr, _Unary):
-        return expr.operator in ("!", "not")
-    if isinstance(expr, _Binary):
-        return expr.operator in _BOOLEAN_RESULT_BINARY_OPERATORS
-    if isinstance(expr, _Call) and isinstance(expr.callee, _Name):
-        return expr.callee.value in _BOOLEAN_RESULT_FUNCTIONS
-    return False
-
-
-def _gml_bool_call(expression: str) -> str:
-    return f"GMRuntime.gml_bool({expression})"
-
-
-def _emit_child(
-    expr: _Expression,
-    parent_precedence: int,
-    is_right_child: bool = False,
-    parent_operator: str | None = None,
-    local_names: Iterable[str] | None = None,
-    scope_context: _ScopeContext | None = None,
-) -> str:
-    text, precedence = _emit_expression(expr, local_names, scope_context=scope_context)
-    needs_parentheses = precedence < parent_precedence
-    if is_right_child and precedence == parent_precedence and parent_operator not in _RIGHT_ASSOCIATIVE:
-        needs_parentheses = True
-    if needs_parentheses:
-        return f"({text})"
-    return text
-
-
-def _strip_comments(source: str) -> str:
-    result: list[str] = []
-    index = 0
-    in_string: str | None = None
-    escaped = False
-    while index < len(source):
-        char = source[index]
-
-        if in_string is not None:
-            result.append(char)
-            if escaped:
-                escaped = False
-            elif char == "\\":
-                escaped = True
-            elif char == in_string:
-                in_string = None
-            index += 1
-            continue
-
-        if char == '"' or char == "'":
-            in_string = char
-            result.append(char)
-            index += 1
-            continue
-
-        if source.startswith("//", index):
-            while index < len(source) and source[index] not in "\r\n":
-                index += 1
-            continue
-
-        if source.startswith("/*", index):
-            end = source.find("*/", index + 2)
-            if end == -1:
-                break
-            index = end + 2
-            continue
-
-        result.append(char)
-        index += 1
-
-    return "".join(result)
-
-
-def _join_macro_continuation_lines(source: str) -> str:
-    lines: list[str] = []
-    pending_macro: str | None = None
-    for line in source.splitlines():
-        current = line if pending_macro is None else f"{pending_macro} {line.lstrip()}"
-        if current.lstrip().startswith("#macro") and current.rstrip().endswith("\\"):
-            pending_macro = current.rstrip()[:-1].rstrip()
-            continue
-        lines.append(current)
-        pending_macro = None
-
-    if pending_macro is not None:
-        lines.append(pending_macro)
-
-    return "\n".join(lines)
-
-
-def _split_statements(source: str) -> list[str]:  # pyright: ignore[reportUnusedFunction]
-    statements: list[str] = []
-    start = 0
-    depth = 0
-    in_string: str | None = None
-    escaped = False
-
-    for index, char in enumerate(source):
-        if in_string is not None:
-            if escaped:
-                escaped = False
-            elif char == "\\":
-                escaped = True
-            elif char == in_string:
-                in_string = None
-            continue
-
-        if char == '"' or char == "'":
-            in_string = char
-            continue
-        if char in "([{":
-            depth += 1
-            continue
-        if char in ")]}" and depth > 0:
-            depth -= 1
-            continue
-        if char == ";" and depth == 0:
-            statements.append(source[start:index])
-            start = index + 1
-
-    trailing = source[start:].strip()
-    if trailing:
-        statements.append(trailing)
-    return [statement for statement in statements if statement.strip()]
-
-
-def _transpile_statement(
-    statement: str,
-    local_names: MutableSet[str] | None = None,
-    instance_variables: MutableSet[str] | None = None,
-    loop_depth: int = 0,
-    continue_depth: int = 0,
-    return_depth: int = 0,
-    finally_depth: int = 0,
-    enum_values: MutableMapping[str, dict[str, int]] | None = None,
-    enum_names: Iterable[str] | None = None,
-    scope_context: _ScopeContext | None = None,
-    inherited_event_call: str | None = None,
-    macro_values: Mapping[str, str] | None = None,
-    generated_counter: list[int] | None = None,
-) -> list[str]:
-    if not statement:
-        return []
-
-    if local_names is None:
-        local_names = set()
-    scope_context = _normalize_scope_context(scope_context)
-    macro_values = macro_values or {}
-    generated_counter = generated_counter if generated_counter is not None else [0]
-
-    if statement == "return":
-        _reject_finally_control_flow(finally_depth)
-        if return_depth <= 0:
-            raise GMLTranspileError("return used outside a function or method")
-        return ["return"]
-    if statement.startswith("return "):
-        _reject_finally_control_flow(finally_depth)
-        if return_depth <= 0:
-            raise GMLTranspileError("return used outside a function or method")
-        return_value = transpile_gml_expression(
-            statement[7:].strip(),
-            local_names,
-            enum_values,
-            enum_names,
-            scope_context=scope_context,
-            macro_values=macro_values,
-        )
-        return [f"return {return_value}"]
-    if statement == "break":
-        _reject_finally_control_flow(finally_depth)
-        if loop_depth <= 0:
-            raise GMLTranspileError("break used outside a loop")
-        return ["break"]
-    if statement == "continue":
-        _reject_finally_control_flow(finally_depth)
-        if continue_depth <= 0:
-            raise GMLTranspileError("continue used outside a loop")
-        return ["continue"]
-    if statement == "exit":
-        _reject_finally_control_flow(finally_depth)
-        return ["return"]
-    event_inherited_lines = _transpile_event_inherited_statement(statement, inherited_event_call)
-    if event_inherited_lines is not None:
-        return event_inherited_lines
-    if statement == "throw":
-        raise GMLTranspileError("throw requires an expression")
-    if statement.startswith("throw "):
-        thrown_value = transpile_gml_expression(
-            statement[6:].strip(),
-            local_names,
-            enum_values,
-            enum_names,
-            scope_context=scope_context,
-            macro_values=macro_values,
-        )
-        return [f"return GMRuntime.gml_throw({thrown_value})"]
-
-    if statement.startswith("delete "):
-        target_source = statement[7:].strip()
-        _reject_constant_assignment_target_name(target_source, macro_values.keys())
-        target_expr = _parse_gml_expression(
-            target_source,
-            enum_values,
-            enum_names,
-            macro_values=macro_values,
-            scope_context=scope_context,
-        )
-        _reject_enum_assignment_target(target_expr, enum_names)
-        _reject_readonly_builtin_assignment_target(target_expr, local_names)
-        if isinstance(target_expr, _Name):
-            _reject_asset_identifier_name(target_expr.value, scope_context)
-        static_target = _static_scope_assignment_parts(target_expr, scope_context)
-        if static_target is not None:
-            static_scope, member_name = static_target
-            return [f"GMRuntime.gml_struct_set({static_scope}, {member_name}, GMRuntime.gml_undefined())"]
-        global_target = _global_scope_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context,
-        )
-        if global_target is not None:
-            global_scope, member_name = global_target
-            return [f"GMRuntime.gml_struct_set({global_scope}, {member_name}, GMRuntime.gml_undefined())"]
-        if not isinstance(target_expr, _Name):
-            raise GMLTranspileError("delete can only be used with variables")
-        scoped_target = _scoped_instance_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context,
-        )
-        if scoped_target is not None:
-            instance_target, member_name = scoped_target
-            return [
-                "GMRuntime.gml_variable_instance_set("
-                f"{instance_target}, {member_name}, GMRuntime.gml_undefined())"
-            ]
-        target = _emit_expression(target_expr, local_names, scope_context=scope_context)[0]
-        return [f"{target} = GMRuntime.gml_undefined()"]
-
-    if statement.startswith("var "):
-        return _transpile_var_statement(
-            statement[4:].strip(),
-            local_names,
-            enum_values,
-            enum_names,
-            scope_context=scope_context,
-            macro_values=macro_values,
-        )
-
-    increment = _parse_increment_statement(statement)
-    if increment is not None:
-        target, delta = increment
-        helper = "gml_add" if delta > 0 else "gml_sub"
-        _reject_constant_assignment_target_name(target, macro_values.keys())
-        target_expr = _parse_gml_expression(
-            target,
-            enum_values,
-            enum_names,
-            macro_values=macro_values,
-            scope_context=scope_context,
-        )
-        if not _is_increment_target_expression(target_expr):
-            raise GMLTranspileError("Increment target must be assignable")
-        _reject_enum_assignment_target(target_expr, enum_names)
-        _reject_readonly_builtin_assignment_target(target_expr, local_names)
-        if isinstance(target_expr, _Name):
-            _reject_asset_identifier_name(target_expr.value, scope_context)
-        static_target = _static_scope_assignment_parts(target_expr, scope_context)
-        if static_target is not None:
-            static_scope, member_name = static_target
-            current_value = f"GMRuntime.gml_struct_get({static_scope}, {member_name})"
-            return [
-                f"GMRuntime.gml_struct_set({static_scope}, {member_name}, "
-                f"GMRuntime.{helper}({current_value}, 1))"
-            ]
-        global_target = _global_scope_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context,
-        )
-        if global_target is not None:
-            global_scope, member_name = global_target
-            current_value = f"GMRuntime.gml_struct_get({global_scope}, {member_name})"
-            return [
-                f"GMRuntime.gml_struct_set({global_scope}, {member_name}, "
-                f"GMRuntime.{helper}({current_value}, 1))"
-            ]
-        scoped_target = _scoped_instance_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context,
-        )
-        if scoped_target is not None:
-            instance_target, member_name = scoped_target
-            current_value = f"GMRuntime.gml_variable_instance_get({instance_target}, {member_name})"
-            return [
-                "GMRuntime.gml_variable_instance_set("
-                f"{instance_target}, {member_name}, "
-                f"GMRuntime.{helper}({current_value}, 1))"
-            ]
-        if isinstance(target_expr, _Index):
-            container = _emit_expression(
-                target_expr.target,
-                local_names,
-                scope_context=scope_context,
-            )[0]
-            index = _emit_expression(
-                target_expr.index,
-                local_names,
-                scope_context=scope_context,
-            )[0]
-            prelude_lines: list[str] = []
-            container = _cache_assignment_part(
-                prelude_lines,
-                target_expr.target,
-                container,
-                generated_counter,
-                "_gml_array_target",
-            )
-            index = _cache_assignment_part(
-                prelude_lines,
-                target_expr.index,
-                index,
-                generated_counter,
-                "_gml_array_index",
-            )
-            current_value = f"GMRuntime.gml_array_get({container}, {index})"
-            return [
-                *prelude_lines,
-                f"GMRuntime.gml_array_set({container}, {index}, "
-                f"GMRuntime.{helper}({current_value}, 1))",
-            ]
-        struct_target = _struct_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context=scope_context,
-        )
-        if struct_target is not None:
-            container, key = struct_target
-            prelude_lines: list[str] = []
-            if isinstance(target_expr, _StructAccess):
-                container = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.target,
-                    container,
-                    generated_counter,
-                    "_gml_struct_target",
-                )
-                key = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.key,
-                    key,
-                    generated_counter,
-                    "_gml_struct_key",
-                )
-            elif isinstance(target_expr, _Member):
-                container = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.target,
-                    container,
-                    generated_counter,
-                    "_gml_struct_target",
-                )
-            current_value = f"GMRuntime.gml_struct_get({container}, {key})"
-            return [
-                *prelude_lines,
-                f"GMRuntime.gml_struct_set({container}, {key}, "
-                f"GMRuntime.{helper}({current_value}, 1))"
-            ]
-        ds_map_target = _ds_map_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context=scope_context,
-        )
-        if ds_map_target is not None:
-            container, key = ds_map_target
-            prelude_lines = []
-            if isinstance(target_expr, _DSMapAccess):
-                container = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.target,
-                    container,
-                    generated_counter,
-                    "_gml_map_target",
-                )
-                key = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.key,
-                    key,
-                    generated_counter,
-                    "_gml_map_key",
-                )
-            current_value = f"GMRuntime.gml_ds_map_find_value({container}, {key})"
-            return [
-                *prelude_lines,
-                f"GMRuntime.gml_ds_map_set({container}, {key}, "
-                f"GMRuntime.{helper}({current_value}, 1))"
-            ]
-        target = _emit_expression(target_expr, local_names, scope_context=scope_context)[0]
-        return [f"{target} = GMRuntime.{helper}({target}, 1)"]
-
-    assignment = _split_assignment(statement)
-    if assignment is not None:
-        target, operator, value = assignment
-        if _split_assignment(value) is not None:
-            raise GMLTranspileError("Chained assignments are not supported")
-        increment_value = _parse_increment_expression(value)
-        if operator in ("=", ":=") and increment_value is not None:
-            increment_target, increment_delta, increment_mode = increment_value
-            increment_expr = _parse_gml_expression(
-                increment_target,
-                enum_values,
-                enum_names,
-                macro_values=macro_values,
-                scope_context=scope_context,
-            )
-            if not _is_increment_target_expression(increment_expr):
-                raise GMLTranspileError("Increment expression target must be assignable")
-            increment_value_text = _emit_expression(
-                increment_expr,
-                local_names,
-                scope_context=scope_context,
-            )[0]
-            prelude_lines: list[str] = []
-            assigned_value = increment_value_text
-            if increment_mode == "postfix":
-                assigned_value = _next_generated_name_from_counter(
-                    generated_counter,
-                    "_gml_increment_value",
-                )
-                prelude_lines.append(f"var {assigned_value} = {increment_value_text}")
-                local_names.add(assigned_value)
-            suffix = "++" if increment_delta > 0 else "--"
-            increment_lines = _transpile_statement(
-                f"{increment_target}{suffix}",
-                local_names,
-                instance_variables,
-                loop_depth=loop_depth,
-                continue_depth=continue_depth,
-                return_depth=return_depth,
-                finally_depth=finally_depth,
-                enum_values=enum_values,
-                enum_names=enum_names,
-                scope_context=scope_context,
-                inherited_event_call=inherited_event_call,
-                macro_values=macro_values,
-                generated_counter=generated_counter,
-            )
-            assignment_lines = _transpile_assignment_to_emitted_value(
-                target,
-                assigned_value,
-                local_names,
-                instance_variables,
-                enum_values=enum_values,
-                enum_names=enum_names,
-                scope_context=scope_context,
-                macro_values=macro_values,
-            )
-            return [*prelude_lines, *increment_lines, *assignment_lines]
-        _reject_constant_assignment_target_name(target, macro_values.keys())
-        target_expr = _parse_gml_expression(
-            target,
-            enum_values,
-            enum_names,
-            macro_values=macro_values,
-            scope_context=scope_context,
-        )
-        _reject_enum_assignment_target(target_expr, enum_names)
-        _reject_readonly_builtin_assignment_target(target_expr, local_names)
-        if isinstance(target_expr, _Name):
-            _reject_asset_identifier_name(target_expr.value, scope_context)
-        static_target = _static_scope_assignment_parts(target_expr, scope_context)
-        global_target = _global_scope_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context,
-        )
-        scoped_target = _scoped_instance_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context,
-        )
-        value = transpile_gml_expression(
-            value,
-            local_names,
-            enum_values,
-            enum_names,
-            scope_context=scope_context,
-            macro_values=macro_values,
-        )
-        if static_target is not None:
-            static_scope, member_name = static_target
-            if operator in ("=", ":="):
-                return [f"GMRuntime.gml_struct_set({static_scope}, {member_name}, {value})"]
-            current_value = f"GMRuntime.gml_struct_get({static_scope}, {member_name})"
-            if operator == "??=":
-                return [
-                    f"if GMRuntime.gml_is_nullish({current_value}):",
-                    f"\tGMRuntime.gml_struct_set({static_scope}, {member_name}, {value})",
-                ]
-            if operator in _COMPOUND_RUNTIME_FUNCTIONS:
-                helper = _COMPOUND_RUNTIME_FUNCTIONS[operator]
-                return [
-                    f"GMRuntime.gml_struct_set({static_scope}, {member_name}, "
-                    f"GMRuntime.{helper}({current_value}, {value}))"
-                ]
-            raise GMLTranspileError("Unsupported static assignment operator")
-        if global_target is not None:
-            global_scope, member_name = global_target
-            if operator in ("=", ":="):
-                return [f"GMRuntime.gml_struct_set({global_scope}, {member_name}, {value})"]
-            current_value = f"GMRuntime.gml_struct_get({global_scope}, {member_name})"
-            if operator == "??=":
-                return [
-                    f"if GMRuntime.gml_is_nullish({current_value}):",
-                    f"\tGMRuntime.gml_struct_set({global_scope}, {member_name}, {value})",
-                ]
-            if operator in _COMPOUND_RUNTIME_FUNCTIONS:
-                helper = _COMPOUND_RUNTIME_FUNCTIONS[operator]
-                return [
-                    f"GMRuntime.gml_struct_set({global_scope}, {member_name}, "
-                    f"GMRuntime.{helper}({current_value}, {value}))"
-                ]
-            raise GMLTranspileError("Unsupported global assignment operator")
-        if scoped_target is not None:
-            instance_target, member_name = scoped_target
-            if operator in ("=", ":="):
-                return [
-                    "GMRuntime.gml_variable_instance_set("
-                    f"{instance_target}, {member_name}, {value})"
-                ]
-            current_value = f"GMRuntime.gml_variable_instance_get({instance_target}, {member_name})"
-            if operator == "??=":
-                return [
-                    f"if GMRuntime.gml_is_nullish({current_value}):",
-                    "\tGMRuntime.gml_variable_instance_set("
-                    f"{instance_target}, {member_name}, {value})",
-                ]
-            if operator in _COMPOUND_RUNTIME_FUNCTIONS:
-                helper = _COMPOUND_RUNTIME_FUNCTIONS[operator]
-                return [
-                    "GMRuntime.gml_variable_instance_set("
-                    f"{instance_target}, {member_name}, "
-                    f"GMRuntime.{helper}({current_value}, {value}))"
-                ]
-            raise GMLTranspileError("Unsupported scoped instance assignment operator")
-        _record_instance_assignment(target, local_names, instance_variables)
-        target = _emit_expression(target_expr, local_names, scope_context=scope_context)[0]
-        if isinstance(target_expr, _Index):
-            container = _emit_expression(
-                target_expr.target,
-                local_names,
-                scope_context=scope_context,
-            )[0]
-            index = _emit_expression(
-                target_expr.index,
-                local_names,
-                scope_context=scope_context,
-            )[0]
-            if operator in ("=", ":="):
-                return [f"GMRuntime.gml_array_set({container}, {index}, {value})"]
-            prelude_lines: list[str] = []
-            container = _cache_assignment_part(
-                prelude_lines,
-                target_expr.target,
-                container,
-                generated_counter,
-                "_gml_array_target",
-            )
-            index = _cache_assignment_part(
-                prelude_lines,
-                target_expr.index,
-                index,
-                generated_counter,
-                "_gml_array_index",
-            )
-            if operator == "??=":
-                current_value = f"GMRuntime.gml_array_get({container}, {index})"
-                return [
-                    *prelude_lines,
-                    f"if GMRuntime.gml_is_nullish({current_value}):",
-                    f"\tGMRuntime.gml_array_set({container}, {index}, {value})",
-                ]
-            if operator in _COMPOUND_RUNTIME_FUNCTIONS:
-                helper = _COMPOUND_RUNTIME_FUNCTIONS[operator]
-                current_value = f"GMRuntime.gml_array_get({container}, {index})"
-                return [
-                    *prelude_lines,
-                    f"GMRuntime.gml_array_set({container}, {index}, "
-                    f"GMRuntime.{helper}({current_value}, {value}))"
-                ]
-        ds_map_target = _ds_map_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context=scope_context,
-        )
-        if ds_map_target is not None:
-            container, key = ds_map_target
-            if operator in ("=", ":="):
-                return [f"GMRuntime.gml_ds_map_set({container}, {key}, {value})"]
-            prelude_lines: list[str] = []
-            if isinstance(target_expr, _DSMapAccess):
-                container = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.target,
-                    container,
-                    generated_counter,
-                    "_gml_map_target",
-                )
-                key = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.key,
-                    key,
-                    generated_counter,
-                    "_gml_map_key",
-                )
-            if operator == "??=":
-                current_value = f"GMRuntime.gml_ds_map_find_value({container}, {key})"
-                return [
-                    *prelude_lines,
-                    f"if GMRuntime.gml_is_nullish({current_value}):",
-                    f"\tGMRuntime.gml_ds_map_set({container}, {key}, {value})",
-                ]
-            if operator in _COMPOUND_RUNTIME_FUNCTIONS:
-                helper = _COMPOUND_RUNTIME_FUNCTIONS[operator]
-                current_value = f"GMRuntime.gml_ds_map_find_value({container}, {key})"
-                return [
-                    *prelude_lines,
-                    f"GMRuntime.gml_ds_map_set({container}, {key}, "
-                    f"GMRuntime.{helper}({current_value}, {value}))"
-                ]
-            raise GMLTranspileError("Unsupported DS map accessor assignment operator")
-        struct_target = _struct_assignment_parts(
-            target_expr,
-            local_names,
-            scope_context=scope_context,
-        )
-        if struct_target is not None:
-            container, key = struct_target
-            if operator in ("=", ":="):
-                return [f"GMRuntime.gml_struct_set({container}, {key}, {value})"]
-            prelude_lines = []
-            if isinstance(target_expr, _StructAccess):
-                container = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.target,
-                    container,
-                    generated_counter,
-                    "_gml_struct_target",
-                )
-                key = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.key,
-                    key,
-                    generated_counter,
-                    "_gml_struct_key",
-                )
-            elif isinstance(target_expr, _Member):
-                container = _cache_assignment_part(
-                    prelude_lines,
-                    target_expr.target,
-                    container,
-                    generated_counter,
-                    "_gml_struct_target",
-                )
-            if operator == "??=":
-                current_value = f"GMRuntime.gml_struct_get({container}, {key})"
-                return [
-                    *prelude_lines,
-                    f"if GMRuntime.gml_is_nullish({current_value}):",
-                    f"\tGMRuntime.gml_struct_set({container}, {key}, {value})",
-                ]
-            if operator in _COMPOUND_RUNTIME_FUNCTIONS:
-                helper = _COMPOUND_RUNTIME_FUNCTIONS[operator]
-                current_value = f"GMRuntime.gml_struct_get({container}, {key})"
-                return [
-                    *prelude_lines,
-                    f"GMRuntime.gml_struct_set({container}, {key}, "
-                    f"GMRuntime.{helper}({current_value}, {value}))"
-                ]
-            raise GMLTranspileError("Unsupported struct member assignment operator")
-        if operator == "??=":
-            return [f"if GMRuntime.gml_is_nullish({target}):", f"\t{target} = {value}"]
-        if operator in _COMPOUND_RUNTIME_FUNCTIONS:
-            return [f"{target} = GMRuntime.{_COMPOUND_RUNTIME_FUNCTIONS[operator]}({target}, {value})"]
-        if operator == ":=":
-            return [f"{target} = {value}"]
-        return [f"{target} {operator} {value}"]
-
-    return [
-        transpile_gml_expression(
-            statement,
-            local_names,
-            enum_values,
-            enum_names,
-            scope_context=scope_context,
-            macro_values=macro_values,
-        )
-    ]
-
-
-def _reject_finally_control_flow(finally_depth: int) -> None:
-    if finally_depth > 0:
-        raise GMLTranspileError("break, continue, exit, and return are not allowed inside finally")
-
-
-def _transpile_event_inherited_statement(
-    statement: str,
-    inherited_event_call: str | None,
-) -> list[str] | None:
-    if not statement.strip().startswith("event_inherited"):
-        return None
-
-    expr = _parse_gml_expression(statement)
-    expr = _unwrap_grouped_expression(expr)
-    if not (
-        isinstance(expr, _Call)
-        and isinstance(expr.callee, _Name)
-        and expr.callee.value == "event_inherited"
-    ):
-        return None
-
-    if expr.args:
-        raise GMLTranspileError("event_inherited does not accept arguments")
-
-    if inherited_event_call is None:
-        return ["pass"]
-    return [inherited_event_call]
-
-
-def _struct_assignment_parts(
-    target_expr: _Expression,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> tuple[str, str] | None:
-    scope_context = _normalize_scope_context(scope_context)
-    if isinstance(target_expr, _StructAccess):
-        container = _emit_expression(
-            target_expr.target,
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        key = _emit_expression(
-            target_expr.key,
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        return container, key
-    if isinstance(target_expr, _Member) and not _uses_direct_member_access(
-        target_expr,
-        scope_context=scope_context,
-    ):
-        container = _emit_expression(
-            target_expr.target,
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        return container, json.dumps(target_expr.member)
-    return None
-
-
-def _ds_map_assignment_parts(
-    target_expr: _Expression,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> tuple[str, str] | None:
-    scope_context = _normalize_scope_context(scope_context)
-    if isinstance(target_expr, _DSMapAccess):
-        container = _emit_expression(
-            target_expr.target,
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        key = _emit_expression(
-            target_expr.key,
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        return container, key
-    return None
-
-
-def _scoped_instance_assignment_parts(
-    target_expr: _Expression,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> tuple[str, str] | None:
-    scope_context = _normalize_scope_context(scope_context)
-    if scope_context.instance_target is None or not isinstance(target_expr, _Name):
-        return None
-
-    name = target_expr.value
-    if (
-        name in local_names
-        or name in _GML_LITERAL_IDENTIFIERS
-        or name in _BUILTIN_ARRAY_VARIABLES
-        or name in _BUILTIN_GLOBAL_VARIABLES
-        or not _is_plain_identifier(name)
-    ):
-        return None
-    return scope_context.instance_target, json.dumps(name)
-
-
-def _static_scope_assignment_parts(
-    target_expr: _Expression,
-    scope_context: _ScopeContext | None = None,
-) -> tuple[str, str] | None:
-    scope_context = _normalize_scope_context(scope_context)
-    if scope_context.static_scope is None or not isinstance(target_expr, _Name):
-        return None
-    name = target_expr.value
-    if name not in scope_context.static_names:
-        return None
-    return scope_context.static_scope, json.dumps(name)
-
-
-def _global_scope_assignment_parts(
-    target_expr: _Expression,
-    local_names: Iterable[str],
-    scope_context: _ScopeContext | None = None,
-) -> tuple[str, str] | None:
-    scope_context = _normalize_scope_context(scope_context)
-    if not isinstance(target_expr, _Name):
-        return None
-    name = target_expr.value
-    if not _name_resolves_to_global(name, local_names, scope_context):
-        return None
-    return "GMRuntime.gml_global_scope()", json.dumps(name)
-
-
-def _record_instance_assignment(
-    target: str,
-    local_names: Iterable[str],
-    instance_variables: MutableSet[str] | None,
-) -> None:
-    if instance_variables is None:
-        return
-
-    tokens = _expression_tokens(target.strip())
-    if len(tokens) != 2 or tokens[0].kind != "IDENT" or tokens[1].kind != "EOF":
-        return
-
-    name = tokens[0].value
-    if name in local_names or name in _BUILTIN_INSTANCE_VARIABLES:
-        return
-    instance_variables.add(name)
-
-
-def _transpile_var_statement(
-    statement: str,
-    local_names: MutableSet[str] | None = None,
-    enum_values: MutableMapping[str, dict[str, int]] | None = None,
-    enum_names: Iterable[str] | None = None,
-    scope_context: _ScopeContext | None = None,
-    macro_values: Mapping[str, str] | None = None,
-) -> list[str]:
-    lines: list[str] = []
-    if local_names is None:
-        local_names = set()
-    scope_context = _normalize_scope_context(scope_context)
-    enum_name_set = frozenset(enum_names or [])
-    macro_values = macro_values or {}
-    for declaration in _split_top_level(statement, ","):
-        declaration = declaration.strip()
-        if not declaration:
-            continue
-        assignment = _split_assignment(declaration)
-        if assignment is None:
-            name = declaration.strip()
-            _validate_gml_identifier(name)
-            _reject_asset_identifier_name(name, scope_context)
-            _reject_constant_declaration_name(name, macro_values.keys())
-            if name in enum_name_set:
-                raise GMLTranspileError("Cannot redeclare enum")
-            lines.append(f"var {_sanitize_gdscript_identifier(name)} = GMRuntime.gml_undefined()")
-            local_names.add(name)
-            continue
-        name, operator, value = assignment
-        if operator not in ("=", ":="):
-            raise GMLTranspileError("Variable declarations only support simple assignments")
-        name = name.strip()
-        _validate_gml_identifier(name)
-        _reject_asset_identifier_name(name, scope_context)
-        _reject_constant_declaration_name(name, macro_values.keys())
-        if name in enum_name_set:
-            raise GMLTranspileError("Cannot redeclare enum")
-        initial_value = transpile_gml_expression(
-            value,
-            local_names,
-            enum_values,
-            enum_names,
-            scope_context=scope_context,
-            macro_values=macro_values,
-        )
-        lines.append(f"var {_sanitize_gdscript_identifier(name)} = {initial_value}")
-        local_names.add(name)
-    return lines
-
-
-def _sanitize_gdscript_identifier(name: str) -> str:
-    if not _is_plain_identifier(name):
-        return name
-    if name in _GDSCRIPT_RESERVED_IDENTIFIERS:
-        return f"{name}_"
-    if name.startswith(_GENERATED_IDENTIFIER_PREFIX):
-        return f"gml_user{name}"
-    return name
-
-
-def _is_plain_identifier(name: str) -> bool:
-    if not name:
-        return False
-    return (name[0].isalpha() or name[0] == "_") and all(
-        char.isalnum() or char == "_" for char in name
-    )
-
-
-def _validate_gml_identifier(name: str) -> None:
-    if not name:
-        raise GMLTranspileError("Expected identifier name")
-    if len(name) > _GML_IDENTIFIER_MAX_LENGTH:
-        raise GMLTranspileError("GML identifier exceeds 64 characters")
-    if not (name[0].isalpha() or name[0] == "_"):
-        raise GMLTranspileError("GML identifier must start with a letter or underscore")
-    if not all(char.isalnum() or char == "_" for char in name):
-        raise GMLTranspileError("GML identifier can only contain letters, numbers, and underscores")
-
-
-def _reject_asset_identifier_name(name: str, scope_context: _ScopeContext) -> None:
-    if name in scope_context.asset_names:
-        raise GMLTranspileError(f"Unscoped identifier '{name}' collides with an asset name")
-
-
-def _parse_increment_statement(statement: str) -> tuple[str, _IncrementDelta] | None:
-    stripped = statement.strip()
-    if stripped.endswith("++"):
-        target = stripped[:-2].strip()
-        if _split_assignment(target) is not None:
-            return None
-        return target, 1
-    if stripped.endswith("--"):
-        target = stripped[:-2].strip()
-        if _split_assignment(target) is not None:
-            return None
-        return target, -1
-    if stripped.startswith("++"):
-        target = stripped[2:].strip()
-        if _split_assignment(target) is not None:
-            return None
-        return target, 1
-    if stripped.startswith("--"):
-        target = stripped[2:].strip()
-        if _split_assignment(target) is not None:
-            return None
-        return target, -1
-    return None
-
-
-def _parse_increment_expression(statement: str) -> tuple[str, _IncrementDelta, _IncrementMode] | None:
-    stripped = statement.strip()
-    if stripped.endswith("++"):
-        target = stripped[:-2].strip()
-        if _split_assignment(target) is not None:
-            return None
-        return target, 1, "postfix"
-    if stripped.endswith("--"):
-        target = stripped[:-2].strip()
-        if _split_assignment(target) is not None:
-            return None
-        return target, -1, "postfix"
-    if stripped.startswith("++"):
-        target = stripped[2:].strip()
-        if _split_assignment(target) is not None:
-            return None
-        return target, 1, "prefix"
-    if stripped.startswith("--"):
-        target = stripped[2:].strip()
-        if _split_assignment(target) is not None:
-            return None
-        return target, -1, "prefix"
-    return None
-
-
-def _is_increment_target_expression(expr: _Expression) -> bool:
-    return isinstance(expr, _Name | _Index | _Member | _StructAccess | _DSMapAccess)
-
-
-def _transpile_assignment_to_emitted_value(
-    target: str,
-    value: str,
-    local_names: MutableSet[str],
-    instance_variables: MutableSet[str] | None,
-    enum_values: MutableMapping[str, dict[str, int]] | None = None,
-    enum_names: Iterable[str] | None = None,
-    scope_context: _ScopeContext | None = None,
-    macro_values: Mapping[str, str] | None = None,
-) -> list[str]:
-    scope_context = _normalize_scope_context(scope_context)
-    macro_values = macro_values or {}
-    _reject_constant_assignment_target_name(target, macro_values.keys())
-    target_expr = _parse_gml_expression(
-        target,
-        enum_values,
-        enum_names,
-        macro_values=macro_values,
-        scope_context=scope_context,
-    )
-    _reject_enum_assignment_target(target_expr, enum_names)
-    _reject_readonly_builtin_assignment_target(target_expr, local_names)
-    if isinstance(target_expr, _Name):
-        _reject_asset_identifier_name(target_expr.value, scope_context)
-    static_target = _static_scope_assignment_parts(target_expr, scope_context)
-    if static_target is not None:
-        static_scope, member_name = static_target
-        return [f"GMRuntime.gml_struct_set({static_scope}, {member_name}, {value})"]
-    global_target = _global_scope_assignment_parts(
-        target_expr,
-        local_names,
-        scope_context,
-    )
-    if global_target is not None:
-        global_scope, member_name = global_target
-        return [f"GMRuntime.gml_struct_set({global_scope}, {member_name}, {value})"]
-    scoped_target = _scoped_instance_assignment_parts(
-        target_expr,
-        local_names,
-        scope_context,
-    )
-    if scoped_target is not None:
-        instance_target, member_name = scoped_target
-        return [
-            "GMRuntime.gml_variable_instance_set("
-            f"{instance_target}, {member_name}, {value})"
-        ]
-    _record_instance_assignment(target, local_names, instance_variables)
-    if isinstance(target_expr, _Index):
-        container = _emit_expression(
-            target_expr.target,
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        index = _emit_expression(
-            target_expr.index,
-            local_names,
-            scope_context=scope_context,
-        )[0]
-        return [f"GMRuntime.gml_array_set({container}, {index}, {value})"]
-    ds_map_target = _ds_map_assignment_parts(
-        target_expr,
-        local_names,
-        scope_context=scope_context,
-    )
-    if ds_map_target is not None:
-        container, key = ds_map_target
-        return [f"GMRuntime.gml_ds_map_set({container}, {key}, {value})"]
-    struct_target = _struct_assignment_parts(
-        target_expr,
-        local_names,
-        scope_context=scope_context,
-    )
-    if struct_target is not None:
-        container, key = struct_target
-        return [f"GMRuntime.gml_struct_set({container}, {key}, {value})"]
-    emitted_target = _emit_expression(
-        target_expr,
-        local_names,
-        scope_context=scope_context,
-    )[0]
-    return [f"{emitted_target} = {value}"]
-
-
-def _split_assignment(statement: str) -> tuple[str, _AssignmentOperator, str] | None:
-    depth = 0
-    in_string: str | None = None
-    escaped = False
-    index = 0
-
-    while index < len(statement):
-        char = statement[index]
-        if in_string is not None:
-            if escaped:
-                escaped = False
-            elif char == "\\":
-                escaped = True
-            elif char == in_string:
-                in_string = None
-            index += 1
-            continue
-
-        if char == '"' or char == "'":
-            in_string = char
-            index += 1
-            continue
-        if char in "([{":
-            depth += 1
-            index += 1
-            continue
-        if char in ")]}" and depth > 0:
-            depth -= 1
-            index += 1
-            continue
-
-        if depth == 0:
-            for operator in _ASSIGNMENT_OPERATORS:
-                if statement.startswith(operator, index):
-                    if operator == "=" and _is_comparison_assignment_false_positive(statement, index):
-                        continue
-                    left = statement[:index].strip()
-                    right = statement[index + len(operator):].strip()
-                    if left and right:
-                        return left, operator, right
-        index += 1
-    return None
-
-
-def _is_comparison_assignment_false_positive(statement: str, index: int) -> bool:
-    previous_char = statement[index - 1] if index > 0 else ""
-    next_char = statement[index + 1] if index + 1 < len(statement) else ""
-    return previous_char in "!<>=?" or next_char == "="
-
-
-def _split_top_level(source: str, separator: str) -> list[str]:
-    parts: list[str] = []
-    start = 0
-    depth = 0
-    in_string: str | None = None
-    escaped = False
-
-    for index, char in enumerate(source):
-        if in_string is not None:
-            if escaped:
-                escaped = False
-            elif char == "\\":
-                escaped = True
-            elif char == in_string:
-                in_string = None
-            continue
-
-        if char == '"' or char == "'":
-            in_string = char
-            continue
-        if char in "([{":
-            depth += 1
-            continue
-        if char in ")]}" and depth > 0:
-            depth -= 1
-            continue
-        if char == separator and depth == 0:
-            parts.append(source[start:index])
-            start = index + 1
-
-    parts.append(source[start:])
-    return parts
