+++ src/conversion/gml_transpiler_parts/__init__.py
@@ -1,0 +1,1 @@
+"""Modular implementation details for src.conversion.gml_transpiler."""
